Fitted model without covariates

The class stores the transaction data and various optimization outputs and options. It is created already when model fitting is initiated and is then used to perform no covariate specific steps during the estimation process. Serves as parent classes to fitted transaction and spending models.

Created with an existing clv.data and clv.model object (or subclasses thereof).

# S4 method for clv.fitted
show(object)

Slots

call

Single language of the call used to create the object

clv.model

Single object of (sub-) class clv.model that determines model-specific behavior.

clv.data

Single object of (sub-) class clv.data that contains the data and temporal information to fit the model to.

prediction.params.model

Numeric vector of the model parameters, set and used solely when predicting. Named after model parameters in original scale and derived from coef().

optimx.estimation.output

A single object of class optimx as returned from method optimx::optimx after optimizing the log-likelihood fitting the model.

optimx.hessian

Single matrix that is the hessian extracted from the last row of the optimization output stored in the slot optimx.estimation.output.

See also