Gamma-Gamma: Log-Likelihood Function

Calculates the Log-Likelihood value for the Gamma-Gamma model.

gg_LL(vLogparams, vX, vM_x)

Arguments

vLogparams

a vector containing the log of the parameters p, q, gamma

vX

frequency vector of length n counting the numbers of purchases

vM_x

the observed average spending for every customer during the calibration time.

Value

Returns the Log-Likelihood value for the Gamma-Gamma model.

Details

vLogparams is a vector with the parameters for the Gamma-Gamma model. It has three parameters (p, q, gamma). The scale parameter for each transaction is distributed across customers according to a gamma distribution with parameters q (shape) and gamma (scale).

References

Colombo R, Jiang W (1999). “A stochastic RFM model.” Journal of Interactive Marketing, 13(3), 2-12.

Fader PS, Hardie BG, Lee K (2005). “RFM and CLV: Using Iso-Value Curves for Customer Base Analysis.” Journal of Marketing Research, 42(4), 415-430.

Fader PS, Hardie BG (2013). “The Gamma-Gamma Model of Monetary Value.” URL http://www.brucehardie.com/notes/025/gamma_gamma.pdf.