Gamma-Gompertz/NBD model

Fits Gamma-Gompertz/NBD models on transactional data with static and without covariates.

# S4 method for class 'clv.data'
ggomnbd(
  clv.data,
  start.params.model = c(),
  optimx.args = list(),
  verbose = TRUE,
  ...
)

# S4 method for class 'clv.data.static.covariates'
ggomnbd(
  clv.data,
  start.params.model = c(),
  optimx.args = list(),
  verbose = TRUE,
  names.cov.life = c(),
  names.cov.trans = c(),
  start.params.life = c(),
  start.params.trans = c(),
  names.cov.constr = c(),
  start.params.constr = c(),
  reg.lambdas = c(),
  ...
)

Arguments

clv.data

The data object on which the model is fitted.

start.params.model

Named start parameters containing the optimization start parameters for the model without covariates.

optimx.args

Additional arguments to control the optimization which are forwarded to optimx::optimx. If multiple optimization methods are specified, only the result of the last method is further processed.

verbose

Show details about the running of the function.

...

Ignored

names.cov.life

Which of the set Lifetime covariates should be used. Missing parameter indicates all covariates shall be used.

names.cov.trans

Which of the set Transaction covariates should be used. Missing parameter indicates all covariates shall be used.

start.params.life

Named start parameters containing the optimization start parameters for all lifetime covariates.

start.params.trans

Named start parameters containing the optimization start parameters for all transaction covariates.

names.cov.constr

Which covariates should be forced to use the same parameters for the lifetime and transaction process. The covariates need to be present as both, lifetime and transaction covariates.

start.params.constr

Named start parameters containing the optimization start parameters for the constraint covariates.

reg.lambdas

Named lambda parameters used for the L2 regularization of the lifetime and the transaction covariate parameters. Lambdas have to be >= 0.

Value

Depending on the data object on which the model was fit, ggomnbd returns either an object of class clv.ggomnbd or clv.ggomnbd.static.cov.

The function summary can be used to obtain and print a summary of the results. The generic accessor functions coefficients, vcov, fitted, logLik, AIC, BIC, and nobs are available.

Details

Model parameters for the GGompertz/NBD model are r, alpha, beta, b and s.
r: shape parameter of the Gamma distribution of the purchase process. The smaller r, the stronger the heterogeneity of the purchase process.
alpha: scale parameter of the Gamma distribution of the purchase process.
beta: scale parameter for the Gamma distribution for the lifetime process.
b: scale parameter of the Gompertz distribution (constant across customers).
s: shape parameter of the Gamma distribution for the lifetime process. The smaller s, the stronger the heterogeneity of customer lifetimes.

If no start parameters are given, r=0.5, alpha=2, b=0.1, s=1, beta=0.1 is used. All model start parameters are required to be > 0. If no start values are given for the covariate parameters, 0.1 is used.

Note that the DERT expression has not been derived (yet) and it consequently is not possible to calculated values for DERT and CLV.

The Gamma-Gompertz/NBD model

There are two key differences of the gamma/Gompertz/NBD (GGompertz/NBD) model compared to the relative to the well-known Pareto/NBD model: (i) its probability density function can exhibit a mode at zero or an interior mode, and (ii) it can be skewed to the right or to the left. Therefore, the GGompertz/NBD model is more flexible than the Pareto/NBD model. According to Bemmaor and Glady (2012) can indicate substantial differences in expected residual lifetimes compared to the Pareto/NBD. The GGompertz/NBD tends to be appropriate when firms are reputed and their offerings are differentiated.

References

Bemmaor AC, Glady N (2012). “Modeling Purchasing Behavior with Sudden “Death”: A Flexible Customer Lifetime Model” Management Science, 58(5), 1012-1021.

Adler J (2022). “Comment on “Modeling Purchasing Behavior with Sudden “Death”: A Flexible Customer Lifetime Model” Management Science 69(3):1929-1930.

The expression for the PMF was derived by Adler J (2024). (unpublished)

See also

Examples

# \donttest{
data("apparelTrans")
clv.data.apparel <- clvdata(apparelTrans, date.format = "ymd",
                            time.unit = "w", estimation.split = 40)

# Fit standard ggomnbd model
ggomnbd(clv.data.apparel)
#> Starting estimation...
#> Estimation finished!
#> GGompertz/NBD Standard Model
#> 
#> Call:
#> ggomnbd(clv.data = clv.data.apparel)
#> 
#> Coefficients:
#>         r      alpha          b          s       beta  
#> 7.866e-01  5.335e+00  5.486e-07  3.569e-01  6.369e-06  
#> KKT1: TRUE 
#> KKT2: FALSE 

# Give initial guesses for the model parameters
ggomnbd(clv.data.apparel,
     start.params.model = c(r=0.5, alpha=15, b=5, beta=10, s=0.5))
#> Starting estimation...
#> Estimation finished!
#> GGompertz/NBD Standard Model
#> 
#> Call:
#> ggomnbd(clv.data = clv.data.apparel, start.params.model = c(r = 0.5, 
#>     alpha = 15, b = 5, beta = 10, s = 0.5))
#> 
#> Coefficients:
#>        r     alpha         b         s      beta  
#> 0.694369  4.918964  9.017822  0.001522  0.628796  
#> KKT1: TRUE 
#> KKT2: FALSE 

# pass additional parameters to the optimizer (optimx)
#    Use Nelder-Mead as optimization method and print
#    detailed information about the optimization process
apparel.ggomnbd <- ggomnbd(clv.data.apparel,
                     optimx.args = list(method="Nelder-Mead",
                                        control=list(trace=6)))
#> Starting estimation...
#> fn is  fn1 
#> Looking for method =  Nelder-Mead 
#> Methods to be used:[1] "Nelder-Mead"
#> optcfg:$fname
#> [1] "fn1"
#> 
#> $npar
#> [1] 5
#> 
#> $ctrl
#> $ctrl$follow.on
#> [1] FALSE
#> 
#> $ctrl$save.failures
#> [1] TRUE
#> 
#> $ctrl$trace
#> [1] 6
#> 
#> $ctrl$kkt
#> [1] TRUE
#> 
#> $ctrl$all.methods
#> [1] FALSE
#> 
#> $ctrl$starttests
#> [1] FALSE
#> 
#> $ctrl$maximize
#> [1] FALSE
#> 
#> $ctrl$dowarn
#> [1] TRUE
#> 
#> $ctrl$usenumDeriv
#> [1] FALSE
#> 
#> $ctrl$kkttol
#> [1] 0.001
#> 
#> $ctrl$kkt2tol
#> [1] 1e-06
#> 
#> $ctrl$badval
#> [1] 8.988466e+307
#> 
#> $ctrl$scaletol
#> [1] 3
#> 
#> $ctrl$have.bounds
#> [1] FALSE
#> 
#> 
#> $usenumDeriv
#> [1] FALSE
#> 
#> $ufn
#> function (par) 
#> fn(par, ...)
#> <bytecode: 0x12e143cb8>
#> <environment: 0x12cffb8a0>
#> 
#> $have.bounds
#> [1] FALSE
#> 
#> $method
#> [1] "Nelder-Mead"
#> 
#> Method:  Nelder-Mead 
#>   Nelder-Mead direct search function minimizer
#> function value for initial parameters = 3451.383126
#>   Scaled convergence tolerance is 5.14296e-05
#> Stepsize computed as 0.230259
#> BUILD              6 3645.413347 3416.553688
#> EXTENSION          8 3565.390170 3204.288840
#> LO-REDUCTION      10 3451.383126 3204.288840
#> LO-REDUCTION      12 3449.118736 3204.288840
#> EXTENSION         14 3444.599205 3129.350248
#> EXTENSION         16 3416.553688 2998.500055
#> LO-REDUCTION      18 3245.918902 2998.500055
#> LO-REDUCTION      20 3243.683531 2998.500055
#> EXTENSION         22 3204.288840 2955.234793
#> LO-REDUCTION      24 3129.350248 2955.234793
#> EXTENSION         26 3001.905548 2884.620697
#> LO-REDUCTION      28 3000.026135 2884.620697
#> LO-REDUCTION      30 2998.500055 2884.620697
#> LO-REDUCTION      32 2959.232462 2884.620697
#> LO-REDUCTION      34 2955.234793 2884.620697
#> LO-REDUCTION      36 2927.505038 2884.620697
#> REFLECTION        38 2902.427323 2884.315233
#> REFLECTION        40 2901.525096 2883.837264
#> LO-REDUCTION      42 2897.157703 2883.837264
#> LO-REDUCTION      44 2889.216744 2883.837264
#> LO-REDUCTION      46 2888.530027 2883.837264
#> LO-REDUCTION      48 2884.944736 2883.008245
#> LO-REDUCTION      50 2884.620697 2882.800641
#> HI-REDUCTION      52 2884.315233 2882.800641
#> EXTENSION         54 2884.237378 2881.208107
#> LO-REDUCTION      56 2883.837264 2881.208107
#> HI-REDUCTION      58 2883.011911 2881.208107
#> LO-REDUCTION      60 2883.008245 2881.208107
#> HI-REDUCTION      62 2882.800641 2881.208107
#> REFLECTION        64 2882.485370 2881.011565
#> LO-REDUCTION      66 2882.092085 2881.011565
#> LO-REDUCTION      68 2882.055175 2881.011565
#> REFLECTION        70 2881.309160 2880.839710
#> REFLECTION        72 2881.208107 2880.782828
#> LO-REDUCTION      74 2881.089791 2880.755613
#> LO-REDUCTION      76 2881.028987 2880.755613
#> LO-REDUCTION      78 2881.011565 2880.755613
#> LO-REDUCTION      80 2880.858578 2880.732309
#> LO-REDUCTION      82 2880.839710 2880.732309
#> REFLECTION        84 2880.834829 2880.706254
#> REFLECTION        86 2880.782828 2880.696556
#> REFLECTION        88 2880.755613 2880.667615
#> HI-REDUCTION      90 2880.747666 2880.667615
#> LO-REDUCTION      92 2880.732309 2880.661230
#> HI-REDUCTION      94 2880.706254 2880.661230
#> LO-REDUCTION      96 2880.696556 2880.657593
#> HI-REDUCTION      98 2880.693930 2880.657593
#> EXTENSION        100 2880.675913 2880.618786
#> LO-REDUCTION     102 2880.667615 2880.618786
#> LO-REDUCTION     104 2880.665994 2880.618786
#> LO-REDUCTION     106 2880.661230 2880.618786
#> EXTENSION        108 2880.657593 2880.604465
#> LO-REDUCTION     110 2880.651656 2880.604465
#> LO-REDUCTION     112 2880.641286 2880.604465
#> LO-REDUCTION     114 2880.632833 2880.604465
#> LO-REDUCTION     116 2880.618786 2880.604465
#> LO-REDUCTION     118 2880.618516 2880.604465
#> LO-REDUCTION     120 2880.612778 2880.604345
#> EXTENSION        122 2880.608029 2880.596413
#> EXTENSION        124 2880.606438 2880.593134
#> LO-REDUCTION     126 2880.606081 2880.593134
#> EXTENSION        128 2880.604465 2880.577619
#> LO-REDUCTION     130 2880.604345 2880.577619
#> EXTENSION        132 2880.596413 2880.555411
#> LO-REDUCTION     134 2880.594303 2880.555411
#> LO-REDUCTION     136 2880.593134 2880.555411
#> EXTENSION        138 2880.580322 2880.534554
#> LO-REDUCTION     140 2880.577619 2880.534554
#> EXTENSION        142 2880.574496 2880.505915
#> LO-REDUCTION     144 2880.559914 2880.505915
#> LO-REDUCTION     146 2880.555411 2880.505915
#> LO-REDUCTION     148 2880.544817 2880.505915
#> LO-REDUCTION     150 2880.534554 2880.505915
#> LO-REDUCTION     152 2880.528350 2880.505915
#> EXTENSION        154 2880.515453 2880.472218
#> LO-REDUCTION     156 2880.509823 2880.472218
#> LO-REDUCTION     158 2880.508563 2880.472218
#> LO-REDUCTION     160 2880.508225 2880.472218
#> EXTENSION        162 2880.505915 2880.446447
#> EXTENSION        164 2880.483130 2880.429632
#> EXTENSION        166 2880.482648 2880.396784
#> LO-REDUCTION     168 2880.474610 2880.396784
#> EXTENSION        170 2880.472218 2880.358319
#> EXTENSION        172 2880.446447 2880.343411
#> LO-REDUCTION     174 2880.429632 2880.343411
#> EXTENSION        176 2880.419037 2880.320851
#> REFLECTION       178 2880.396784 2880.291119
#> LO-REDUCTION     180 2880.367532 2880.291119
#> LO-REDUCTION     182 2880.358319 2880.291119
#> LO-REDUCTION     184 2880.343411 2880.291119
#> EXTENSION        186 2880.337270 2880.284009
#> REFLECTION       188 2880.320851 2880.273325
#> LO-REDUCTION     190 2880.320619 2880.273325
#> EXTENSION        192 2880.308562 2880.230326
#> EXTENSION        194 2880.291119 2880.222281
#> LO-REDUCTION     196 2880.284009 2880.222281
#> EXTENSION        198 2880.278345 2880.207193
#> LO-REDUCTION     200 2880.273325 2880.207193
#> LO-REDUCTION     202 2880.247505 2880.207193
#> LO-REDUCTION     204 2880.230326 2880.207193
#> LO-REDUCTION     206 2880.222281 2880.204934
#> HI-REDUCTION     208 2880.217388 2880.204934
#> LO-REDUCTION     210 2880.212599 2880.204934
#> REFLECTION       212 2880.210725 2880.200148
#> EXTENSION        214 2880.209634 2880.191243
#> LO-REDUCTION     216 2880.207193 2880.191243
#> LO-REDUCTION     218 2880.205132 2880.191243
#> LO-REDUCTION     220 2880.204934 2880.191243
#> EXTENSION        222 2880.200148 2880.184330
#> EXTENSION        224 2880.197430 2880.182832
#> LO-REDUCTION     226 2880.196981 2880.182832
#> REFLECTION       228 2880.191440 2880.178881
#> EXTENSION        230 2880.191243 2880.169137
#> EXTENSION        232 2880.184330 2880.163988
#> HI-REDUCTION     234 2880.183635 2880.163988
#> LO-REDUCTION     236 2880.182832 2880.163988
#> HI-REDUCTION     238 2880.178881 2880.163988
#> EXTENSION        240 2880.173469 2880.152182
#> LO-REDUCTION     242 2880.171248 2880.152182
#> LO-REDUCTION     244 2880.169137 2880.152182
#> LO-REDUCTION     246 2880.165847 2880.152182
#> LO-REDUCTION     248 2880.163988 2880.152182
#> LO-REDUCTION     250 2880.162214 2880.152182
#> EXTENSION        252 2880.160404 2880.144087
#> LO-REDUCTION     254 2880.160355 2880.144087
#> LO-REDUCTION     256 2880.155917 2880.144087
#> EXTENSION        258 2880.154387 2880.134935
#> LO-REDUCTION     260 2880.152182 2880.134935
#> EXTENSION        262 2880.150152 2880.129203
#> LO-REDUCTION     264 2880.146650 2880.129203
#> EXTENSION        266 2880.144087 2880.113605
#> LO-REDUCTION     268 2880.137967 2880.113605
#> EXTENSION        270 2880.134935 2880.105485
#> LO-REDUCTION     272 2880.131133 2880.105485
#> LO-REDUCTION     274 2880.129203 2880.105485
#> REFLECTION       276 2880.120754 2880.100703
#> REFLECTION       278 2880.113605 2880.098911
#> HI-REDUCTION     280 2880.113199 2880.098911
#> EXTENSION        282 2880.106776 2880.092752
#> LO-REDUCTION     284 2880.105645 2880.092752
#> REFLECTION       286 2880.105485 2880.091314
#> LO-REDUCTION     288 2880.100703 2880.091314
#> LO-REDUCTION     290 2880.098911 2880.091314
#> REFLECTION       292 2880.094935 2880.088628
#> LO-REDUCTION     294 2880.092752 2880.088628
#> HI-REDUCTION     296 2880.092481 2880.088628
#> REFLECTION       298 2880.091935 2880.088188
#> EXTENSION        300 2880.091314 2880.086547
#> HI-REDUCTION     302 2880.090597 2880.086547
#> EXTENSION        304 2880.090592 2880.084388
#> EXTENSION        306 2880.088628 2880.079570
#> LO-REDUCTION     308 2880.088616 2880.079570
#> LO-REDUCTION     310 2880.088188 2880.079570
#> LO-REDUCTION     312 2880.086547 2880.079570
#> EXTENSION        314 2880.084388 2880.073241
#> EXTENSION        316 2880.081362 2880.068855
#> LO-REDUCTION     318 2880.079930 2880.068855
#> LO-REDUCTION     320 2880.079886 2880.068855
#> EXTENSION        322 2880.079570 2880.061506
#> EXTENSION        324 2880.073771 2880.055536
#> EXTENSION        326 2880.073241 2880.045161
#> LO-REDUCTION     328 2880.070304 2880.045161
#> EXTENSION        330 2880.068855 2880.027420
#> LO-REDUCTION     332 2880.061506 2880.027420
#> LO-REDUCTION     334 2880.055536 2880.027420
#> EXTENSION        336 2880.051058 2880.019707
#> EXTENSION        338 2880.045161 2880.013869
#> LO-REDUCTION     340 2880.041506 2880.013869
#> EXTENSION        342 2880.035306 2879.991036
#> LO-REDUCTION     344 2880.027420 2879.991036
#> LO-REDUCTION     346 2880.019707 2879.991036
#> LO-REDUCTION     348 2880.017380 2879.991036
#> EXTENSION        350 2880.013869 2879.983670
#> LO-REDUCTION     352 2880.003012 2879.983670
#> LO-REDUCTION     354 2880.000805 2879.983670
#> EXTENSION        356 2879.999359 2879.974338
#> LO-REDUCTION     358 2879.991036 2879.974338
#> LO-REDUCTION     360 2879.990564 2879.974338
#> LO-REDUCTION     362 2879.986973 2879.974338
#> LO-REDUCTION     364 2879.983670 2879.974338
#> LO-REDUCTION     366 2879.982051 2879.974338
#> EXTENSION        368 2879.979359 2879.968045
#> LO-REDUCTION     370 2879.977452 2879.968045
#> LO-REDUCTION     372 2879.976805 2879.968045
#> EXTENSION        374 2879.976732 2879.964115
#> LO-REDUCTION     376 2879.974338 2879.964115
#> EXTENSION        378 2879.973192 2879.960495
#> LO-REDUCTION     380 2879.972120 2879.960495
#> EXTENSION        382 2879.968498 2879.951311
#> LO-REDUCTION     384 2879.968045 2879.951311
#> EXTENSION        386 2879.964115 2879.944338
#> LO-REDUCTION     388 2879.961163 2879.944338
#> EXTENSION        390 2879.960495 2879.942566
#> REFLECTION       392 2879.951577 2879.937080
#> LO-REDUCTION     394 2879.951311 2879.937080
#> LO-REDUCTION     396 2879.948666 2879.937080
#> LO-REDUCTION     398 2879.944338 2879.937080
#> LO-REDUCTION     400 2879.942991 2879.937080
#> LO-REDUCTION     402 2879.942566 2879.937080
#> LO-REDUCTION     404 2879.941541 2879.937080
#> EXTENSION        406 2879.941290 2879.935043
#> EXTENSION        408 2879.941271 2879.933028
#> EXTENSION        410 2879.938267 2879.926015
#> LO-REDUCTION     412 2879.937380 2879.926015
#> LO-REDUCTION     414 2879.937080 2879.926015
#> LO-REDUCTION     416 2879.935043 2879.926015
#> LO-REDUCTION     418 2879.933028 2879.926015
#> LO-REDUCTION     420 2879.932102 2879.926015
#> EXTENSION        422 2879.928057 2879.917610
#> LO-REDUCTION     424 2879.927416 2879.917610
#> LO-REDUCTION     426 2879.926639 2879.917610
#> LO-REDUCTION     428 2879.926038 2879.917610
#> EXTENSION        430 2879.926015 2879.908942
#> LO-REDUCTION     432 2879.921397 2879.908942
#> LO-REDUCTION     434 2879.920810 2879.908942
#> LO-REDUCTION     436 2879.920049 2879.908942
#> EXTENSION        438 2879.917610 2879.903374
#> EXTENSION        440 2879.915462 2879.893563
#> EXTENSION        442 2879.911807 2879.882909
#> LO-REDUCTION     444 2879.909546 2879.882909
#> LO-REDUCTION     446 2879.908942 2879.882909
#> EXTENSION        448 2879.903374 2879.868798
#> EXTENSION        450 2879.893563 2879.843187
#> LO-REDUCTION     452 2879.887869 2879.843187
#> LO-REDUCTION     454 2879.885893 2879.843187
#> EXTENSION        456 2879.882909 2879.810134
#> LO-REDUCTION     458 2879.868798 2879.810134
#> EXTENSION        460 2879.854104 2879.770188
#> LO-REDUCTION     462 2879.845583 2879.770188
#> LO-REDUCTION     464 2879.843187 2879.770188
#> EXTENSION        466 2879.826529 2879.743037
#> EXTENSION        468 2879.810134 2879.703396
#> LO-REDUCTION     470 2879.785920 2879.703396
#> EXTENSION        472 2879.777056 2879.672095
#> EXTENSION        474 2879.770188 2879.622467
#> LO-REDUCTION     476 2879.743037 2879.622467
#> LO-REDUCTION     478 2879.707449 2879.622467
#> LO-REDUCTION     480 2879.703396 2879.622467
#> LO-REDUCTION     482 2879.672095 2879.622467
#> LO-REDUCTION     484 2879.663901 2879.622467
#> LO-REDUCTION     486 2879.658657 2879.622467
#> LO-REDUCTION     488 2879.646515 2879.622467
#> EXTENSION        490 2879.642766 2879.606911
#> EXTENSION        492 2879.633890 2879.596177
#> EXTENSION        494 2879.628206 2879.576101
#> LO-REDUCTION     496 2879.622739 2879.576101
#> LO-REDUCTION     498 2879.622467 2879.576101
#> LO-REDUCTION     500 2879.606911 2879.576101
#> LO-REDUCTION     502 2879.596177 2879.576101
#> LO-REDUCTION     504 2879.587757 2879.576101
#> LO-REDUCTION     506 2879.587261 2879.576101
#> EXTENSION        508 2879.583796 2879.562375
#> LO-REDUCTION     510 2879.583480 2879.562375
#> LO-REDUCTION     512 2879.579241 2879.562375
#> LO-REDUCTION     514 2879.577072 2879.562375
#> REFLECTION       516 2879.576101 2879.561070
#> EXTENSION        518 2879.572211 2879.550780
#> EXTENSION        520 2879.563829 2879.535893
#> LO-REDUCTION     522 2879.562703 2879.535893
#> LO-REDUCTION     524 2879.562375 2879.535893
#> EXTENSION        526 2879.561070 2879.529823
#> LO-REDUCTION     528 2879.550780 2879.529823
#> EXTENSION        530 2879.544305 2879.515202
#> LO-REDUCTION     532 2879.542593 2879.515202
#> LO-REDUCTION     534 2879.535893 2879.515202
#> LO-REDUCTION     536 2879.534692 2879.515202
#> EXTENSION        538 2879.529823 2879.509514
#> EXTENSION        540 2879.523825 2879.504620
#> EXTENSION        542 2879.519565 2879.499066
#> LO-REDUCTION     544 2879.518944 2879.499066
#> LO-REDUCTION     546 2879.515202 2879.499066
#> REFLECTION       548 2879.509514 2879.498275
#> HI-REDUCTION     550 2879.504620 2879.498275
#> REFLECTION       552 2879.504567 2879.495123
#> EXTENSION        554 2879.501180 2879.489470
#> LO-REDUCTION     556 2879.499441 2879.489470
#> LO-REDUCTION     558 2879.499066 2879.489470
#> LO-REDUCTION     560 2879.498275 2879.489470
#> LO-REDUCTION     562 2879.495123 2879.489470
#> HI-REDUCTION     564 2879.494884 2879.489470
#> EXTENSION        566 2879.492984 2879.486114
#> LO-REDUCTION     568 2879.492146 2879.486114
#> LO-REDUCTION     570 2879.491854 2879.486114
#> REFLECTION       572 2879.490998 2879.484774
#> REFLECTION       574 2879.489470 2879.483990
#> EXTENSION        576 2879.486634 2879.481901
#> LO-REDUCTION     578 2879.486282 2879.481901
#> LO-REDUCTION     580 2879.486114 2879.481901
#> EXTENSION        582 2879.484774 2879.478557
#> LO-REDUCTION     584 2879.483990 2879.478557
#> LO-REDUCTION     586 2879.483194 2879.478557
#> LO-REDUCTION     588 2879.482623 2879.478557
#> LO-REDUCTION     590 2879.481901 2879.478557
#> EXTENSION        592 2879.481433 2879.477880
#> REFLECTION       594 2879.481036 2879.477711
#> REFLECTION       596 2879.479980 2879.476808
#> EXTENSION        598 2879.478638 2879.475912
#> HI-REDUCTION     600 2879.478557 2879.475912
#> REFLECTION       602 2879.477880 2879.475864
#> EXTENSION        604 2879.477711 2879.474392
#> REFLECTION       606 2879.476967 2879.474058
#> LO-REDUCTION     608 2879.476808 2879.474058
#> LO-REDUCTION     610 2879.475912 2879.474058
#> EXTENSION        612 2879.475864 2879.473500
#> REFLECTION       614 2879.475366 2879.473485
#> REFLECTION       616 2879.474646 2879.472725
#> REFLECTION       618 2879.474392 2879.472369
#> EXTENSION        620 2879.474058 2879.471466
#> LO-REDUCTION     622 2879.473500 2879.471466
#> LO-REDUCTION     624 2879.473485 2879.471466
#> LO-REDUCTION     626 2879.472725 2879.471466
#> REFLECTION       628 2879.472369 2879.471032
#> LO-REDUCTION     630 2879.472201 2879.471032
#> LO-REDUCTION     632 2879.471744 2879.470980
#> HI-REDUCTION     634 2879.471514 2879.470980
#> HI-REDUCTION     636 2879.471466 2879.470980
#> LO-REDUCTION     638 2879.471233 2879.470980
#> LO-REDUCTION     640 2879.471193 2879.470980
#> REFLECTION       642 2879.471181 2879.470964
#> EXTENSION        644 2879.471041 2879.470806
#> HI-REDUCTION     646 2879.471032 2879.470806
#> EXTENSION        648 2879.471029 2879.470765
#> REFLECTION       650 2879.470980 2879.470757
#> EXTENSION        652 2879.470964 2879.470567
#> LO-REDUCTION     654 2879.470930 2879.470567
#> LO-REDUCTION     656 2879.470806 2879.470567
#> LO-REDUCTION     658 2879.470765 2879.470567
#> EXTENSION        660 2879.470757 2879.470430
#> LO-REDUCTION     662 2879.470646 2879.470430
#> LO-REDUCTION     664 2879.470646 2879.470430
#> LO-REDUCTION     666 2879.470640 2879.470430
#> LO-REDUCTION     668 2879.470567 2879.470430
#> REFLECTION       670 2879.470509 2879.470415
#> REFLECTION       672 2879.470496 2879.470403
#> Exiting from Nelder Mead minimizer
#>     674 function evaluations used
#> Post processing for method  Nelder-Mead 
#> Successful convergence! 
#> Compute Hessian approximation at finish of  Nelder-Mead 
#> Compute gradient approximation at finish of  Nelder-Mead 
#> Save results from method  Nelder-Mead 
#> $par
#>       log.r   log.alpha       log.b       log.s    log.beta 
#>  -0.2375547   1.6766811 -10.5256723  -1.0348880  -8.0822959 
#> 
#> $value
#> [1] 2879.47
#> 
#> $message
#> NULL
#> 
#> $convcode
#> [1] 0
#> 
#> $fevals
#> function 
#>      674 
#> 
#> $gevals
#> gradient 
#>       NA 
#> 
#> $nitns
#> [1] NA
#> 
#> $kkt1
#> [1] TRUE
#> 
#> $kkt2
#> [1] FALSE
#> 
#> $xtimes
#> user.self 
#>      0.11 
#> 
#> Assemble the answers
#> Estimation finished!

# estimated coefs
coef(apparel.ggomnbd)
#>            r        alpha            b            s         beta 
#> 7.885537e-01 5.347778e+00 2.683852e-05 3.552662e-01 3.089609e-04 

# summary of the fitted model
summary(apparel.ggomnbd)
#> GGompertz/NBD Standard  Model 
#> 
#> Call:
#> ggomnbd(clv.data = clv.data.apparel, optimx.args = list(method = "Nelder-Mead", 
#>     control = list(trace = 6)))
#> 
#> Fitting period:                               
#> Estimation start  2005-01-03   
#> Estimation end    2005-10-10   
#> Estimation length 40.0000 Weeks
#> 
#> Coefficients:
#>        Estimate Std. Error z-val Pr(>|z|)    
#> r     7.886e-01  1.329e-01 5.935 2.93e-09 ***
#> alpha 5.348e+00  9.051e-01 5.908 3.46e-09 ***
#> b     2.684e-05  1.512e-03 0.018   0.9858    
#> s     3.553e-01  1.821e-01 1.951   0.0511 .  
#> beta  3.090e-04  1.740e-02 0.018   0.9858    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Optimization info:                  
#> LL     -2879.4704 
#> AIC    5768.9408  
#> BIC    5786.5481  
#> KKT 1  TRUE       
#> KKT 2  FALSE      
#> fevals 674.0000   
#> Method Nelder-Mead

# predict CLV etc for holdout period
predict(apparel.ggomnbd)
#> Predicting from 2005-10-11 until (incl.) 2006-07-16 (39.86 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length actual.x
#>      <char>       <Date>      <Date>         <num>    <int>
#>   1:      1   2005-10-11  2006-07-16      39.85714        0
#>   2:     10   2005-10-11  2006-07-16      39.85714        0
#>   3:    100   2005-10-11  2006-07-16      39.85714       23
#>   4:   1000   2005-10-11  2006-07-16      39.85714       23
#>   5:   1001   2005-10-11  2006-07-16      39.85714       11
#>  ---                                                       
#> 246:   1219   2005-10-11  2006-07-16      39.85714       14
#> 247:    122   2005-10-11  2006-07-16      39.85714        0
#> 248:   1220   2005-10-11  2006-07-16      39.85714        0
#> 249:   1221   2005-10-11  2006-07-16      39.85714        9
#> 250:   1222   2005-10-11  2006-07-16      39.85714        0
#>      actual.total.spending predicted.mean.spending predicted.total.spending
#>                      <num>                   <num>                    <num>
#>   1:                  0.00                39.95483                  8.84709
#>   2:                  0.00                55.23031                 51.28053
#>   3:                737.53                43.57390                590.38416
#>   4:               1069.91                41.60921                548.33416
#>   5:                364.00                45.58153                161.01553
#>  ---                                                                       
#> 246:                413.76                33.58728                121.33685
#> 247:                  0.00                39.95483                  8.84709
#> 248:                  0.00                39.95483                  8.84709
#> 249:                302.65                34.28958                147.48199
#> 250:                  0.00                47.35500                 27.59058
#>             CET    PAlive
#>           <num>     <num>
#>   1:  0.2214273 0.3565817
#>   2:  0.9284854 0.4228197
#>   3: 13.5490321 0.9157440
#>   4: 13.1781929 0.9967856
#>   5:  3.5324733 0.5104117
#>  ---                     
#> 246:  3.6125836 0.9580159
#> 247:  0.2214273 0.3565817
#> 248:  0.2214273 0.3565817
#> 249:  4.3010730 0.9435522
#> 250:  0.5826328 0.4136682

# predict CLV etc for the next 15 periods
predict(apparel.ggomnbd, prediction.end = 15)
#> Predicting from 2005-10-11 until (incl.) 2006-01-23 (15 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length actual.x
#>      <char>       <Date>      <Date>         <int>    <int>
#>   1:      1   2005-10-11  2006-01-23            15        0
#>   2:     10   2005-10-11  2006-01-23            15        0
#>   3:    100   2005-10-11  2006-01-23            15       10
#>   4:   1000   2005-10-11  2006-01-23            15       10
#>   5:   1001   2005-10-11  2006-01-23            15        6
#>  ---                                                       
#> 246:   1219   2005-10-11  2006-01-23            15       11
#> 247:    122   2005-10-11  2006-01-23            15        0
#> 248:   1220   2005-10-11  2006-01-23            15        0
#> 249:   1221   2005-10-11  2006-01-23            15        7
#> 250:   1222   2005-10-11  2006-01-23            15        0
#>      actual.total.spending predicted.mean.spending predicted.total.spending
#>                      <num>                   <num>                    <num>
#>   1:                  0.00                39.95483                 3.545474
#>   2:                  0.00                55.23031                20.550689
#>   3:                224.71                43.57390               236.596626
#>   4:                343.11                41.60921               219.745076
#>   5:                206.65                45.58153                64.527019
#>  ---                                                                       
#> 246:                325.05                33.58728                48.625778
#> 247:                  0.00                39.95483                 3.545474
#> 248:                  0.00                39.95483                 3.545474
#> 249:                158.46                34.28958                59.103449
#> 250:                  0.00                47.35500                11.056931
#>             CET    PAlive
#>           <num>     <num>
#>   1: 0.08873706 0.3565817
#>   2: 0.37209079 0.4228197
#>   3: 5.42977859 0.9157440
#>   4: 5.28116470 0.9967856
#>   5: 1.41563972 0.5104117
#>  ---                     
#> 246: 1.44774395 0.9580159
#> 247: 0.08873706 0.3565817
#> 248: 0.08873706 0.3565817
#> 249: 1.72365628 0.9435522
#> 250: 0.23349026 0.4136682
# }
# \donttest{
# To estimate the ggomnbd model with static covariates,
#   add static covariates to the data
data("apparelStaticCov")
clv.data.static.cov <-
 SetStaticCovariates(clv.data.apparel,
                     data.cov.life = apparelStaticCov,
                     names.cov.life = c("Gender", "Channel"),
                     data.cov.trans = apparelStaticCov,
                     names.cov.trans = c("Gender", "Channel"))

# Fit ggomnbd with static covariates
ggomnbd(clv.data.static.cov)
#> Starting estimation...
#> Estimation finished!
#> GGompertz/NBD with Static Covariates Model
#> 
#> Call:
#> ggomnbd(clv.data = clv.data.static.cov)
#> 
#> Coefficients:
#>             r          alpha              b              s           beta  
#>       1.42983       35.57287        0.05487        0.15192        0.24793  
#>   life.Gender   life.Channel   trans.Gender  trans.Channel  
#>       2.22640       -2.66256        1.43888        0.37360  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> Constraints:     FALSE 
#> Regularization:  FALSE 

# Give initial guesses for both covariate parameters
ggomnbd(clv.data.static.cov, start.params.trans = c(Gender=0.75, Channel=0.7),
                   start.params.life  = c(Gender=0.5, Channel=0.5))
#> Starting estimation...
#> Estimation finished!
#> GGompertz/NBD with Static Covariates Model
#> 
#> Call:
#> ggomnbd(clv.data = clv.data.static.cov, start.params.life = c(Gender = 0.5, 
#>     Channel = 0.5), start.params.trans = c(Gender = 0.75, Channel = 0.7))
#> 
#> Coefficients:
#>             r          alpha              b              s           beta  
#>       1.42876       35.57571        0.05517        0.15159        0.25168  
#>   life.Gender   life.Channel   trans.Gender  trans.Channel  
#>       2.24021       -2.66604        1.43957        0.37360  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> Constraints:     FALSE 
#> Regularization:  FALSE 

# Use regularization
ggomnbd(clv.data.static.cov, reg.lambdas = c(trans = 5, life=5))
#> Starting estimation...
#> Estimation finished!
#> GGompertz/NBD with Static Covariates Model
#> 
#> Call:
#> ggomnbd(clv.data = clv.data.static.cov, reg.lambdas = c(trans = 5, 
#>     life = 5))
#> 
#> Coefficients:
#>             r          alpha              b              s           beta  
#>     0.7912116      5.4394264      0.0002167      0.3599823      0.0025482  
#>   life.Gender   life.Channel   trans.Gender  trans.Channel  
#>    -0.0005591     -0.0024270      0.0107061      0.0088366  
#> KKT1: TRUE 
#> KKT2: FALSE 
#> Constraints:     FALSE 
#> Regularization:  TRUE 

# Force the same coefficient to be used for both covariates
ggomnbd(clv.data.static.cov, names.cov.constr = "Gender",
                   start.params.constr = c(Gender=0.5))
#> Starting estimation...
#> Estimation finished!
#> GGompertz/NBD with Static Covariates Model
#> 
#> Call:
#> ggomnbd(clv.data = clv.data.static.cov, names.cov.constr = "Gender", 
#>     start.params.constr = c(Gender = 0.5))
#> 
#> Coefficients:
#>             r          alpha              b              s           beta  
#>       1.44983       34.83834        0.03633        0.18033        0.11914  
#>  life.Channel  trans.Channel  constr.Gender  
#>      -2.27039        0.37973        1.39920  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> Constraints:     TRUE 
#> Regularization:  FALSE 

# Fit model only with the Channel covariate for life but
# keep all trans covariates as is
ggomnbd(clv.data.static.cov, names.cov.life = c("Channel"))
#> Starting estimation...
#> Estimation finished!
#> GGompertz/NBD with Static Covariates Model
#> 
#> Call:
#> ggomnbd(clv.data = clv.data.static.cov, names.cov.life = c("Channel"))
#> 
#> Coefficients:
#>             r          alpha              b              s           beta  
#>       1.52864       32.06713        0.09947        0.10148        0.01608  
#>  life.Channel   trans.Gender  trans.Channel  
#>      -3.52120        1.28527        0.35031  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> Constraints:     FALSE 
#> Regularization:  FALSE 
# }