GGompertz/NBD: Log-Likelihood functions

Calculates the Log-Likelihood values for the GGompertz/NBD model with and without covariates.

The function ggomnbd_nocov_LL_ind calculates the individual LogLikelihood values for each customer for the given parameters.

The function ggomnbd_nocov_LL_sum calculates the LogLikelihood value summed across customers for the given parameters.

The function ggomnbd_staticcov_LL_ind calculates the individual LogLikelihood values for each customer for the given parameters and covariates.

The function ggomnbd_staticcov_LL_sum calculates the individual LogLikelihood values summed across customers.

ggomnbd_nocov_LL_ind(vLogparams, vX, vT_x, vT_cal)

ggomnbd_nocov_LL_sum(vLogparams, vX, vT_x, vT_cal)

ggomnbd_staticcov_LL_ind(vParams, vX, vT_x, vT_cal, mCov_life, mCov_trans)

ggomnbd_staticcov_LL_sum(vParams, vX, vT_x, vT_cal, mCov_life, mCov_trans)

Arguments

vLogparams
vector with the GGompertz/NBD model parameters at log scale. See Details.

vX
Frequency vector of length n counting the numbers of purchases.

vT_x
Recency vector of length n.

vT_cal
Vector of length n indicating the total number of periods of observation.

vParams
vector with the parameters for the GGompertz/NBD model at log scale and the static covariates at original scale. See Details.

mCov_life
Matrix containing the covariates data affecting the lifetime process. One column for each covariate.

mCov_trans
Matrix containing the covariates data affecting the transaction process. One column for each covariate.

Value

Returns the respective Log-Likelihood value(s) for the GGompertz/NBD model with or without covariates.

Details

vLogparams is a vector with model parameters r, alpha_0, b, s, beta_0 at log-scale, in this order.

vParams is vector with the GGompertz/NBD model parameters at log scale, followed by the parameters for the lifetime covariates at original scale and then followed by the parameters for the transaction covariates at original scale

mCov_trans is a matrix containing the covariates data of the time-invariant covariates that affect the transaction process. Each column represents a different covariate. For every column a gamma parameter needs to added to vParams at the respective position.

mCov_life is a matrix containing the covariates data of the time-invariant covariates that affect the lifetime process. Each column represents a different covariate. For every column a gamma parameter needs to added to vParams at the respective position.

References

Bemmaor AC, Glady N (2012). “Modeling Purchasing Behavior with Sudden “Death”: A Flexible Customer Lifetime Model” Management Science, 58(5), 1012-1021.