GGompertz/NBD: Probability Mass Function (PMF)

Calculate P(X(t)=x), the probability that a randomly selected customer makes exactly x transactions in the interval (0, t].

ggomnbd_nocov_PMF(r, alpha_0, b, s, beta_0, x, vT_i)

ggomnbd_staticcov_PMF(
  r,
  alpha_0,
  b,
  s,
  beta_0,
  x,
  vCovParams_trans,
  vCovParams_life,
  mCov_life,
  mCov_trans,
  vT_i
)

Arguments

r

shape parameter of the Gamma distribution of the purchase process. The smaller r, the stronger the heterogeneity of the purchase process.

alpha_0

scale parameter of the Gamma distribution of the purchase process.

b

scale parameter of the Gompertz distribution (constant across customers)

s

shape parameter of the Gamma distribution for the lifetime process The smaller s, the stronger the heterogeneity of customer lifetimes.

beta_0

scale parameter for the Gamma distribution for the lifetime process

x

The number of transactions to calculate the probability for (unsigned integer).

vT_i

Number of periods since the customer came alive.

vCovParams_trans

Vector of estimated parameters for the transaction covariates.

vCovParams_life

Vector of estimated parameters for the lifetime covariates.

mCov_life

Matrix containing the covariates data affecting the lifetime process. One column for each covariate.

mCov_trans

Matrix containing the covariates data affecting the transaction process. One column for each covariate.

Value

Returns a vector of probabilities.

Details

mCov_trans is a matrix containing the covariates data of the time-invariant covariates that affect the transaction process. Each column represents a different covariate. For every column a gamma parameter needs to added to vCovParams_trans at the respective position.

mCov_life is a matrix containing the covariates data of the time-invariant covariates that affect the lifetime process. Each column represents a different covariate. For every column a gamma parameter needs to added to vCovParams_life at the respective position.

References

Bemmaor AC, Glady N (2012). “Modeling Purchasing Behavior with Sudden “Death”: A Flexible Customer Lifetime Model” Management Science, 58(5), 1012-1021.

Adler J (2022). “Comment on “Modeling Purchasing Behavior with Sudden “Death”: A Flexible Customer Lifetime Model” Management Science 69(3):1929-1930.

The expression for the PMF was derived by Adler J (2024). (unpublished)