Pareto/NBD models

Fits Pareto/NBD models on transactional data with and without covariates.

# S4 method for class 'clv.data'
pnbd(
  clv.data,
  start.params.model = c(),
  use.cor = FALSE,
  start.param.cor = c(),
  optimx.args = list(),
  verbose = TRUE,
  ...
)

# S4 method for class 'clv.data.static.covariates'
pnbd(
  clv.data,
  start.params.model = c(),
  use.cor = FALSE,
  start.param.cor = c(),
  optimx.args = list(),
  verbose = TRUE,
  names.cov.life = c(),
  names.cov.trans = c(),
  start.params.life = c(),
  start.params.trans = c(),
  names.cov.constr = c(),
  start.params.constr = c(),
  reg.lambdas = c(),
  ...
)

# S4 method for class 'clv.data.dynamic.covariates'
pnbd(
  clv.data,
  start.params.model = c(),
  use.cor = FALSE,
  start.param.cor = c(),
  optimx.args = list(),
  verbose = TRUE,
  names.cov.life = c(),
  names.cov.trans = c(),
  start.params.life = c(),
  start.params.trans = c(),
  names.cov.constr = c(),
  start.params.constr = c(),
  reg.lambdas = c(),
  ...
)

Arguments

clv.data

The data object on which the model is fitted.

start.params.model

Named start parameters containing the optimization start parameters for the model without covariates.

use.cor

Whether the correlation between the transaction and lifetime process should be estimated.

start.param.cor

Start parameter for the optimization of the correlation.

optimx.args

Additional arguments to control the optimization which are forwarded to optimx::optimx. If multiple optimization methods are specified, only the result of the last method is further processed.

verbose

Show details about the running of the function.

...

Ignored

names.cov.life

Which of the set Lifetime covariates should be used. Missing parameter indicates all covariates shall be used.

names.cov.trans

Which of the set Transaction covariates should be used. Missing parameter indicates all covariates shall be used.

start.params.life

Named start parameters containing the optimization start parameters for all lifetime covariates.

start.params.trans

Named start parameters containing the optimization start parameters for all transaction covariates.

names.cov.constr

Which covariates should be forced to use the same parameters for the lifetime and transaction process. The covariates need to be present as both, lifetime and transaction covariates.

start.params.constr

Named start parameters containing the optimization start parameters for the constraint covariates.

reg.lambdas

Named lambda parameters used for the L2 regularization of the lifetime and the transaction covariate parameters. Lambdas have to be >= 0.

Value

Depending on the data object on which the model was fit, pnbd returns either an object of class clv.pnbd, clv.pnbd.static.cov, or clv.pnbd.dynamic.cov.

The function summary can be used to obtain and print a summary of the results. The generic accessor functions coefficients, vcov, fitted, logLik, AIC, BIC, and nobs are available.

Details

Model parameters for the Pareto/NBD model are alpha, r, beta, and s.
s: shape parameter of the Gamma distribution for the lifetime process. The smaller s, the stronger the heterogeneity of customer lifetimes.
beta: rate parameter for the Gamma distribution for the lifetime process.
r: shape parameter of the Gamma distribution of the purchase process. The smaller r, the stronger the heterogeneity of the purchase process.
alpha: rate parameter of the Gamma distribution of the purchase process.

Based on these parameters, the average purchase rate while customers are active is r/alpha and the average dropout rate is s/beta.

Ideally, the starting parameters for r and s represent your best guess concerning the heterogeneity of customers in their buy and die rate. If covariates are included into the model additionally parameters for the covariates affecting the attrition and the purchase process are part of the model.

If no start parameters are given, r=0.5, alpha=15, s=0.5, beta=10 is used for all model parameters and 0.1 for covariate parameters. The model start parameters are required to be > 0.

The Pareto/NBD model

The Pareto/NBD is the first model addressing the issue of modeling customer purchases and attrition simultaneously for non-contractual settings. The model uses a Pareto distribution, a combination of an Exponential and a Gamma distribution, to explicitly model customers' (unobserved) attrition behavior in addition to customers' purchase process.
In general, the Pareto/NBD model consist of two parts. A first process models the purchase behavior of customers as long as the customers are active. A second process models customers' attrition. Customers live (and buy) for a certain unknown time until they become inactive and "die". Customer attrition is unobserved. Inactive customers may not be reactivated. For technical details we refer to the original paper by Schmittlein, Morrison and Colombo (1987) and the detailed technical note of Fader and Hardie (2005).

Pareto/NBD model with static covariates

The standard Pareto/NBD model captures heterogeneity was solely using Gamma distributions. However, often exogenous knowledge, such as for example customer demographics, is available. The supplementary knowledge may explain part of the heterogeneity among the customers and therefore increase the predictive accuracy of the model. In addition, we can rely on these parameter estimates for inference, i.e. identify and quantify effects of contextual factors on the two underlying purchase and attrition processes. For technical details we refer to the technical note by Fader and Hardie (2007).

Pareto/NBD model with dynamic covariates

In many real-world applications customer purchase and attrition behavior may be influenced by covariates that vary over time. In consequence, the timing of a purchase and the corresponding value of at covariate a that time becomes relevant. Time-varying covariates can affect customer on aggregated level as well as on an individual level: In the first case, all customers are affected simultaneously, in the latter case a covariate is only relevant for a particular customer. For technical details we refer to the paper by Bachmann, Meierer and Näf (2020).

Note

The Pareto/NBD model with dynamic covariates can currently not be fit with data that has a temporal resolution of less than one day (data that was built with time unit hours).

References

Schmittlein DC, Morrison DG, Colombo R (1987). “Counting Your Customers: Who-Are They and What Will They Do Next?” Management Science, 33(1), 1-24.

Bachmann P, Meierer M, Naef, J (2021). “The Role of Time-Varying Contextual Factors in Latent Attrition Models for Customer Base Analysis” Marketing Science 40(4). 783-809.

Fader PS, Hardie BGS (2005). “A Note on Deriving the Pareto/NBD Model and Related Expressions.” URL http://www.brucehardie.com/notes/009/pareto_nbd_derivations_2005-11-05.pdf.

Fader PS, Hardie BGS (2007). “Incorporating time-invariant covariates into the Pareto/NBD and BG/NBD models.” URL http://www.brucehardie.com/notes/019/time_invariant_covariates.pdf.

Fader PS, Hardie BGS (2020). “Deriving an Expression for P(X(t)=x) Under the Pareto/NBD Model.” URL https://www.brucehardie.com/notes/012/pareto_NBD_pmf_derivation_rev.pdf

See also

Examples

# \donttest{
data("apparelTrans")
clv.data.apparel <- clvdata(apparelTrans, date.format = "ymd",
                            time.unit = "w", estimation.split = 40)

# Fit standard pnbd model
pnbd(clv.data.apparel)
#> Starting estimation...
#> Estimation finished!
#> Pareto/NBD Standard Model
#> 
#> Call:
#> pnbd(clv.data = clv.data.apparel)
#> 
#> Coefficients:
#>       r    alpha        s     beta  
#>  0.7858   5.3312   0.3606  11.8221  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> 
#> Used Options:
#> Correlation:     FALSE 

# Give initial guesses for the model parameters
pnbd(clv.data.apparel,
     start.params.model = c(r=0.5, alpha=15, s=0.5, beta=10))
#> Starting estimation...
#> Estimation finished!
#> Pareto/NBD Standard Model
#> 
#> Call:
#> pnbd(clv.data = clv.data.apparel, start.params.model = c(r = 0.5, 
#>     alpha = 15, s = 0.5, beta = 10))
#> 
#> Coefficients:
#>       r    alpha        s     beta  
#>  0.7858   5.3312   0.3606  11.8221  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> 
#> Used Options:
#> Correlation:     FALSE 

# pass additional parameters to the optimizer (optimx)
#    Use Nelder-Mead as optimization method and print
#    detailed information about the optimization process
apparel.pnbd <- pnbd(clv.data.apparel,
                     optimx.args = list(method="Nelder-Mead",
                                        control=list(trace=6)))
#> Starting estimation...
#> fn is  fn1 
#> Looking for method =  Nelder-Mead 
#> Methods to be used:[1] "Nelder-Mead"
#> optcfg:$fname
#> [1] "fn1"
#> 
#> $npar
#> [1] 4
#> 
#> $ctrl
#> $ctrl$follow.on
#> [1] FALSE
#> 
#> $ctrl$save.failures
#> [1] TRUE
#> 
#> $ctrl$trace
#> [1] 6
#> 
#> $ctrl$kkt
#> [1] TRUE
#> 
#> $ctrl$all.methods
#> [1] FALSE
#> 
#> $ctrl$starttests
#> [1] FALSE
#> 
#> $ctrl$maximize
#> [1] FALSE
#> 
#> $ctrl$dowarn
#> [1] TRUE
#> 
#> $ctrl$usenumDeriv
#> [1] FALSE
#> 
#> $ctrl$kkttol
#> [1] 0.001
#> 
#> $ctrl$kkt2tol
#> [1] 1e-06
#> 
#> $ctrl$badval
#> [1] 8.988466e+307
#> 
#> $ctrl$scaletol
#> [1] 3
#> 
#> $ctrl$have.bounds
#> [1] FALSE
#> 
#> 
#> $usenumDeriv
#> [1] FALSE
#> 
#> $ufn
#> function (par) 
#> fn(par, ...)
#> <bytecode: 0x12e143cb8>
#> <environment: 0x13a1cd2a8>
#> 
#> $have.bounds
#> [1] FALSE
#> 
#> $method
#> [1] "Nelder-Mead"
#> 
#> Method:  Nelder-Mead 
#>   Nelder-Mead direct search function minimizer
#> function value for initial parameters = 3054.783875
#>   Scaled convergence tolerance is 4.55198e-05
#> Stepsize computed as 0.270805
#> BUILD              5 3130.171249 3012.666615
#> EXTENSION          7 3073.973545 2934.108266
#> LO-REDUCTION       9 3054.783875 2934.108266
#> EXTENSION         11 3046.196662 2890.274544
#> REFLECTION        13 3012.666615 2885.087876
#> REFLECTION        15 2953.224793 2880.665503
#> LO-REDUCTION      17 2934.108266 2880.665503
#> LO-REDUCTION      19 2890.274544 2880.665503
#> HI-REDUCTION      21 2886.242588 2880.665503
#> HI-REDUCTION      23 2885.087876 2880.665503
#> LO-REDUCTION      25 2882.431079 2880.665503
#> LO-REDUCTION      27 2882.373576 2880.286217
#> HI-REDUCTION      29 2881.398381 2880.286217
#> EXTENSION         31 2881.323229 2879.756503
#> LO-REDUCTION      33 2880.665503 2879.748586
#> LO-REDUCTION      35 2880.356248 2879.748586
#> LO-REDUCTION      37 2880.286217 2879.675815
#> HI-REDUCTION      39 2880.004056 2879.675815
#> LO-REDUCTION      41 2879.821861 2879.675815
#> LO-REDUCTION      43 2879.756503 2879.675815
#> LO-REDUCTION      45 2879.748586 2879.669232
#> HI-REDUCTION      47 2879.691775 2879.668339
#> HI-REDUCTION      49 2879.686220 2879.652021
#> HI-REDUCTION      51 2879.675815 2879.652021
#> LO-REDUCTION      53 2879.669232 2879.652021
#> LO-REDUCTION      55 2879.668339 2879.652021
#> HI-REDUCTION      57 2879.658041 2879.652021
#> HI-REDUCTION      59 2879.653671 2879.650385
#> REFLECTION        61 2879.653554 2879.648557
#> REFLECTION        63 2879.653031 2879.645463
#> HI-REDUCTION      65 2879.652021 2879.645463
#> HI-REDUCTION      67 2879.650385 2879.645463
#> EXTENSION         69 2879.648557 2879.639035
#> LO-REDUCTION      71 2879.647520 2879.639035
#> LO-REDUCTION      73 2879.646944 2879.639035
#> EXTENSION         75 2879.645463 2879.631207
#> LO-REDUCTION      77 2879.643620 2879.631207
#> EXTENSION         79 2879.642257 2879.627502
#> LO-REDUCTION      81 2879.639035 2879.627502
#> EXTENSION         83 2879.635380 2879.611646
#> LO-REDUCTION      85 2879.631207 2879.611646
#> LO-REDUCTION      87 2879.629717 2879.611646
#> EXTENSION         89 2879.627502 2879.596842
#> EXTENSION         91 2879.620950 2879.581402
#> EXTENSION         93 2879.614418 2879.564653
#> EXTENSION         95 2879.611646 2879.541865
#> EXTENSION         97 2879.596842 2879.512661
#> EXTENSION         99 2879.581402 2879.484609
#> LO-REDUCTION     101 2879.564653 2879.484609
#> LO-REDUCTION     103 2879.541865 2879.484609
#> LO-REDUCTION     105 2879.512661 2879.484609
#> REFLECTION       107 2879.489001 2879.482623
#> REFLECTION       109 2879.487967 2879.477322
#> REFLECTION       111 2879.485488 2879.473094
#> LO-REDUCTION     113 2879.484609 2879.473094
#> HI-REDUCTION     115 2879.482623 2879.473094
#> LO-REDUCTION     117 2879.477322 2879.473094
#> LO-REDUCTION     119 2879.477157 2879.473094
#> EXTENSION        121 2879.475811 2879.470484
#> HI-REDUCTION     123 2879.474730 2879.470484
#> LO-REDUCTION     125 2879.473882 2879.470484
#> LO-REDUCTION     127 2879.473094 2879.470484
#> LO-REDUCTION     129 2879.472662 2879.470381
#> HI-REDUCTION     131 2879.470911 2879.470381
#> HI-REDUCTION     133 2879.470860 2879.470381
#> REFLECTION       135 2879.470730 2879.470063
#> LO-REDUCTION     137 2879.470492 2879.470063
#> LO-REDUCTION     139 2879.470484 2879.470063
#> HI-REDUCTION     141 2879.470381 2879.470063
#> LO-REDUCTION     143 2879.470174 2879.470039
#> REFLECTION       145 2879.470168 2879.470019
#> REFLECTION       147 2879.470072 2879.469968
#> HI-REDUCTION     149 2879.470063 2879.469968
#> HI-REDUCTION     151 2879.470039 2879.469968
#> HI-REDUCTION     153 2879.470019 2879.469968
#> Exiting from Nelder Mead minimizer
#>     155 function evaluations used
#> Post processing for method  Nelder-Mead 
#> Successful convergence! 
#> Compute Hessian approximation at finish of  Nelder-Mead 
#> Compute gradient approximation at finish of  Nelder-Mead 
#> Save results from method  Nelder-Mead 
#> $par
#>      log.r  log.alpha      log.s   log.beta 
#> -0.2402537  1.6745955 -1.0319390  2.4497084 
#> 
#> $value
#> [1] 2879.47
#> 
#> $message
#> NULL
#> 
#> $convcode
#> [1] 0
#> 
#> $fevals
#> function 
#>      155 
#> 
#> $gevals
#> gradient 
#>       NA 
#> 
#> $nitns
#> [1] NA
#> 
#> $kkt1
#> [1] TRUE
#> 
#> $kkt2
#> [1] TRUE
#> 
#> $xtimes
#> user.self 
#>     0.018 
#> 
#> Assemble the answers
#> Estimation finished!

# estimated coefs
coef(apparel.pnbd)
#>          r      alpha          s       beta 
#>  0.7864283  5.3366358  0.3563154 11.5849676 

# summary of the fitted model
summary(apparel.pnbd)
#> Pareto/NBD Standard  Model 
#> 
#> Call:
#> pnbd(clv.data = clv.data.apparel, optimx.args = list(method = "Nelder-Mead", 
#>     control = list(trace = 6)))
#> 
#> Fitting period:                               
#> Estimation start  2005-01-03   
#> Estimation end    2005-10-10   
#> Estimation length 40.0000 Weeks
#> 
#> Coefficients:
#>       Estimate Std. Error z-val Pr(>|z|)    
#> r       0.7864     0.1323 5.944 2.78e-09 ***
#> alpha   5.3366     0.9025 5.913 3.36e-09 ***
#> s       0.3563     0.1830 1.947   0.0516 .  
#> beta   11.5850    10.6138 1.092   0.2751    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Optimization info:                  
#> LL     -2879.4700 
#> AIC    5766.9399  
#> BIC    5781.0258  
#> KKT 1  TRUE       
#> KKT 2  TRUE       
#> fevals 155.0000   
#> Method Nelder-Mead
#> 
#> Used Options:                 
#> Correlation FALSE

# predict CLV etc for holdout period
predict(apparel.pnbd)
#> Predicting from 2005-10-11 until (incl.) 2006-07-16 (39.86 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length actual.x
#>      <char>       <Date>      <Date>         <num>    <int>
#>   1:      1   2005-10-11  2006-07-16      39.85714        0
#>   2:     10   2005-10-11  2006-07-16      39.85714        0
#>   3:    100   2005-10-11  2006-07-16      39.85714       23
#>   4:   1000   2005-10-11  2006-07-16      39.85714       23
#>   5:   1001   2005-10-11  2006-07-16      39.85714       11
#>  ---                                                       
#> 246:   1219   2005-10-11  2006-07-16      39.85714       14
#> 247:    122   2005-10-11  2006-07-16      39.85714        0
#> 248:   1220   2005-10-11  2006-07-16      39.85714        0
#> 249:   1221   2005-10-11  2006-07-16      39.85714        9
#> 250:   1222   2005-10-11  2006-07-16      39.85714        0
#>      actual.total.spending    PAlive        CET       DERT
#>                      <num>     <num>      <num>      <num>
#>   1:                  0.00 0.3574086  0.2213742 0.06124756
#>   2:                  0.00 0.4229262  0.9281438 0.25678942
#>   3:                737.53 0.9156653 13.5482683 3.74839745
#>   4:               1069.91 0.9967828 13.1783679 3.64605717
#>   5:                364.00 0.5102077  3.5307112 0.97684137
#>  ---                                                      
#> 246:                413.76 0.9579838  3.6113745 0.99915845
#> 247:                  0.00 0.3574086  0.2213742 0.06124756
#> 248:                  0.00 0.3574086  0.2213742 0.06124756
#> 249:                302.65 0.9435096  4.2999136 1.18965648
#> 250:                  0.00 0.4139543  0.5824260 0.16113973
#>      predicted.mean.spending predicted.total.spending predicted.CLV
#>                        <num>                    <num>         <num>
#>   1:                39.95483                 8.844967      2.447136
#>   2:                55.23031                51.261670     14.182559
#>   3:                43.57390               590.350880    163.332293
#>   4:                41.60921               548.341440    151.709548
#>   5:                45.58153               160.935208     44.525921
#>  ---                                                               
#> 246:                33.58728               121.296238     33.559013
#> 247:                39.95483                 8.844967      2.447136
#> 248:                39.95483                 8.844967      2.447136
#> 249:                34.28958               147.442230     40.792821
#> 250:                47.35500                27.580786      7.630772

# predict CLV etc for the next 15 periods
predict(apparel.pnbd, prediction.end = 15)
#> Predicting from 2005-10-11 until (incl.) 2006-01-23 (15 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length actual.x
#>      <char>       <Date>      <Date>         <int>    <int>
#>   1:      1   2005-10-11  2006-01-23            15        0
#>   2:     10   2005-10-11  2006-01-23            15        0
#>   3:    100   2005-10-11  2006-01-23            15       10
#>   4:   1000   2005-10-11  2006-01-23            15       10
#>   5:   1001   2005-10-11  2006-01-23            15        6
#>  ---                                                       
#> 246:   1219   2005-10-11  2006-01-23            15       11
#> 247:    122   2005-10-11  2006-01-23            15        0
#> 248:   1220   2005-10-11  2006-01-23            15        0
#> 249:   1221   2005-10-11  2006-01-23            15        7
#> 250:   1222   2005-10-11  2006-01-23            15        0
#>      actual.total.spending    PAlive        CET       DERT
#>                      <num>     <num>      <num>      <num>
#>   1:                  0.00 0.3574086 0.08872108 0.06124756
#>   2:                  0.00 0.4229262 0.37197621 0.25678942
#>   3:                224.71 0.9156653 5.42979800 3.74839745
#>   4:                343.11 0.9967828 5.28155143 3.64605717
#>   5:                206.65 0.5102077 1.41501839 0.97684137
#>  ---                                                      
#> 246:                325.05 0.9579838 1.44734614 0.99915845
#> 247:                  0.00 0.3574086 0.08872108 0.06124756
#> 248:                  0.00 0.3574086 0.08872108 0.06124756
#> 249:                158.46 0.9435096 1.72329495 1.18965648
#> 250:                  0.00 0.4139543 0.23342140 0.16113973
#>      predicted.mean.spending predicted.total.spending predicted.CLV
#>                        <num>                    <num>         <num>
#>   1:                39.95483                 3.544835      2.447136
#>   2:                55.23031                20.544361     14.182559
#>   3:                43.57390               236.597472    163.332293
#>   4:                41.60921               219.761167    151.709548
#>   5:                45.58153                64.498698     44.525921
#>  ---                                                               
#> 246:                33.58728                48.612417     33.559013
#> 247:                39.95483                 3.544835      2.447136
#> 248:                39.95483                 3.544835      2.447136
#> 249:                34.28958                59.091059     40.792821
#> 250:                47.35500                11.053671      7.630772
# }
# \donttest{
# Estimate correlation as well
pnbd(clv.data.apparel, use.cor = TRUE)
#> Starting estimation...
#> Estimation finished!
#> Pareto/NBD Standard Model
#> 
#> Call:
#> pnbd(clv.data = clv.data.apparel, use.cor = TRUE)
#> 
#> Coefficients:
#>               r            alpha                s             beta  
#>         0.78443          5.31381          0.35410         11.52172  
#> Cor(life,trans)  
#>         0.00495  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> 
#> Used Options:
#> Correlation:     TRUE 
# }
# \donttest{
# To estimate the pnbd model with static covariates,
#   add static covariates to the data
data("apparelStaticCov")
clv.data.static.cov <-
 SetStaticCovariates(clv.data.apparel,
                     data.cov.life = apparelStaticCov,
                     names.cov.life = c("Gender", "Channel"),
                     data.cov.trans = apparelStaticCov,
                     names.cov.trans = c("Gender", "Channel"))

# Fit pnbd with static covariates
pnbd(clv.data.static.cov)
#> Starting estimation...
#> Estimation finished!
#> Pareto/NBD with Static Covariates Model
#> 
#> Call:
#> pnbd(clv.data = clv.data.static.cov)
#> 
#> Coefficients:
#>             r          alpha              s           beta    life.Gender  
#>        1.4201        35.6425         0.2729         8.5634         1.5215  
#>  life.Channel   trans.Gender  trans.Channel  
#>       -1.7010         1.4228         0.4024  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> 
#> Used Options:
#> Correlation:     FALSE 
#> Constraints:     FALSE 
#> Regularization:  FALSE 

# Give initial guesses for both covariate parameters
pnbd(clv.data.static.cov, start.params.trans = c(Gender=0.75, Channel=0.7),
                   start.params.life  = c(Gender=0.5, Channel=0.5))
#> Starting estimation...
#> Estimation finished!
#> Pareto/NBD with Static Covariates Model
#> 
#> Call:
#> pnbd(clv.data = clv.data.static.cov, start.params.life = c(Gender = 0.5, 
#>     Channel = 0.5), start.params.trans = c(Gender = 0.75, Channel = 0.7))
#> 
#> Coefficients:
#>             r          alpha              s           beta    life.Gender  
#>        1.4175        35.6357         0.2730         8.6759         1.5335  
#>  life.Channel   trans.Gender  trans.Channel  
#>       -1.7026         1.4240         0.4028  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> 
#> Used Options:
#> Correlation:     FALSE 
#> Constraints:     FALSE 
#> Regularization:  FALSE 

# Use regularization
pnbd(clv.data.static.cov, reg.lambdas = c(trans = 5, life=5))
#> Starting estimation...
#> Estimation finished!
#> Pareto/NBD with Static Covariates Model
#> 
#> Call:
#> pnbd(clv.data = clv.data.static.cov, reg.lambdas = c(trans = 5, 
#>     life = 5))
#> 
#> Coefficients:
#>             r          alpha              s           beta    life.Gender  
#>     0.7921514      5.4472757      0.3603584     11.7402269     -0.0004837  
#>  life.Channel   trans.Gender  trans.Channel  
#>    -0.0023797      0.0106719      0.0087680  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> 
#> Used Options:
#> Correlation:     FALSE 
#> Constraints:     FALSE 
#> Regularization:  TRUE 

# Force the same coefficient to be used for both covariates
pnbd(clv.data.static.cov, names.cov.constr = "Gender",
                   start.params.constr = c(Gender=0.5))
#> Starting estimation...
#> Estimation finished!
#> Pareto/NBD with Static Covariates Model
#> 
#> Call:
#> pnbd(clv.data = clv.data.static.cov, names.cov.constr = "Gender", 
#>     start.params.constr = c(Gender = 0.5))
#> 
#> Coefficients:
#>             r          alpha              s           beta   life.Channel  
#>        1.4230        35.4434         0.2741         7.8762        -1.6881  
#> trans.Channel  constr.Gender  
#>        0.4017         1.4156  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> 
#> Used Options:
#> Correlation:     FALSE 
#> Constraints:     TRUE 
#> Regularization:  FALSE 

# Fit model only with the Channel covariate for life but
# keep all trans covariates as is
pnbd(clv.data.static.cov, names.cov.life = c("Channel"))
#> Starting estimation...
#> Estimation finished!
#> Pareto/NBD with Static Covariates Model
#> 
#> Call:
#> pnbd(clv.data = clv.data.static.cov, names.cov.life = c("Channel"))
#> 
#> Coefficients:
#>             r          alpha              s           beta   life.Channel  
#>        1.4924        32.1618         0.2496         1.7968        -1.6954  
#>  trans.Gender  trans.Channel  
#>        1.2723         0.3933  
#> KKT1: TRUE 
#> KKT2: TRUE 
#> 
#> Used Options:
#> Correlation:     FALSE 
#> Constraints:     FALSE 
#> Regularization:  FALSE 
# }
# Add dynamic covariates data to the data object
#   add dynamic covariates to the data
# \donttest{
if (FALSE) { # \dontrun{
data("apparelDynCov")
clv.data.dyn.cov <-
  SetDynamicCovariates(clv.data = clv.data.apparel,
                       data.cov.life = apparelDynCov,
                       data.cov.trans = apparelDynCov,
                       names.cov.life = c("Marketing", "Gender", "Channel"),
                       names.cov.trans = c("Marketing", "Gender", "Channel"),
                       name.date = "Cov.Date")

# Fit PNBD with dynamic covariates
pnbd(clv.data.dyn.cov)

# The same fitting options as for the
#  static covariate are available
pnbd(clv.data.dyn.cov, reg.lambdas = c(trans=10, life=2))
} # }
# }