Predict CLV from a fitted model

Probabilistic customer attrition models predict in general three expected characteristics for every customer:

"conditional expected transactions" (CET), which is the number of transactions to expect form a customer during the prediction period,

"probability of a customer being alive" (PAlive) at the end of the estimation period and

"discounted expected residual transactions" (DERT) for every customer, which is the total number of transactions for the residual lifetime of a customer discounted to the end of the estimation period. In the case of time-varying covariates, instead of DERT, "discounted expected conditional transactions" (DECT) is predicted. DECT does only cover a finite time horizon in contrast to DERT. For continuous.discount.factor=0, DECT corresponds to CET.

If spending information was provided in the clvdata object, by default a Gamma/Gamma model is fitted to predict spending and calculate the predicted CLV. In this case, the prediction additionally contains the following two columns:

predicted spending estimated by a Gamma/Gamma model

the customer lifetime value (CLV).

# S3 method for clv.fitted
predict(
  object,
  newdata = NULL,
  prediction.end = NULL,
  predict.spending = clv.data.has.spending(object@clv.data),
  continuous.discount.factor = 0.1,
  verbose = TRUE,
  ...
)

# S4 method for clv.fitted
predict(
  object,
  newdata = NULL,
  prediction.end = NULL,
  predict.spending = clv.data.has.spending(object@clv.data),
  continuous.discount.factor = 0.1,
  verbose = TRUE,
  ...
)

Arguments

object
A fitted clv model for which prediction is desired.

newdata
A clv data object for which predictions should be made with the fitted model. If none or NULL is given, predictions are made for the data on which the model was fit.

prediction.end
Until what point in time to predict. This can be the number of periods (numeric) or a form of date/time object. See details.

predict.spending
Whether the spending and CLV should be calculated and reported additionally. Only possible if the transaction data contains spending information.

continuous.discount.factor
continuous discount factor to use

verbose
Show details about the running of the function.

...
Ignored

Value

An object of class data.table with each columns containing the predictions:

Id

The respective customer identifier

period.first

First timepoint of prediction period

period.last

Last timepoint of prediction period

period.length

Number of time units covered by the period indicated by period.first and period.last (including both ends).

PAlive

Probability to be alive at the end of the estimation period

CET

The Conditional Expected Transactions

DERT or DECT

Discounted Expected Residual Transactions or Discounted Expected Conditional Transactions for dynamic covariates models

actual.x

Actual number of transactions until prediction.end. Only if there is a holdout period and the prediction ends in it.

actual.Spending

Actual spending until prediction.end. Only if there is a holdout period and the prediction ends in it, 0 otherwise.

predicted.Spending

The spending as predicted by the Gamma-Gamma model.

predicted.CLV

Customer Lifetime Value based on DERT and predicted spending.

Details

The newdata argument has to be a clv data object of the exact same class as the data object on which the model was fit. In case the model was fit with covariates, newdata needs to contain identically named covariate data.

The use case for newdata is mainly two-fold: First, to estimate model parameters only on a sample of the data and then use the fitted model object to predict or plot for the full data set provided through newdata. Second, for models with dynamic covariates, to provide a clv data object with longer covariates than contained in the data on which the model was estimated what allows to predict or plot further. When providing newdata, some models might require additional steps that can significantly increase runtime.

prediction.end indicates until when to predict or plot and can be given as either a point in time (of class Date, POSIXct, or character) or the number of periods. If prediction.end is of class character, the date/time format set when creating the data object is used for parsing. If prediction.end is the number of periods, the end of the fitting period serves as the reference point from which periods are counted. Only full periods may be specified. If prediction.end is omitted or NULL, it defaults to the end of the holdout period if present and to the end of the estimation period otherwise.

The first prediction period is defined to start right after the end of the estimation period. If for example weekly time units are used and the estimation period ends on Sunday 2019-01-01, then the first day of the first prediction period is Monday 2019-01-02. Each prediction period includes a total of 7 days and the first prediction period therefore will end on, and include, Sunday 2019-01-08. Subsequent prediction periods again start on Mondays and end on Sundays. If prediction.end indicates a timepoint on which to end, this timepoint is included in the prediction period.

predict.spending uses a Gamma/Gamma model to predict customer spending. This option is only available if customer spending information was provided when the data object was created.

continuous.discount.factor allows to adjust the discount rate used to estimated the discounted expected transactions (DERT). The default value is 0.1 (=10%). Note that a continuous rate needs to be provided.

The Gamma-Gamma model to Predict Spending

Most probabilistic latent customer attrition model capture future customer behavior as a combination of the customer's purchase and attrition process. However, in order to derive a monetary value such as CLV, customer spending also has to be considered. To model customer spending the Gamma/Gamma is a popular choice.

References

Schmittlein DC, Morrison DG, Colombo R (1987). “Counting Your Customers: Who-Are They and What Will They Do Next?” Management Science, 33(1), 1–24.

Fader PS, Hardie BGS (2005). “A Note on Deriving the Pareto/NBD Model and Related Expressions.” URL http://www.brucehardie.com/notes/009/pareto_nbd_derivations_2005-11-05.pdf.

Fader PS, Hardie BG (2007). "Incorporating time-invariant covariates into the Pareto/NBD and BG/NBD models." URL http://www.brucehardie.com/notes/019/time_invariant_covariates.pdf.

Colombo R, Jiang W (1999). "A stochastic RFM model." Journal of Interactive Marketing, 13(3), 2–12.

Fader PS, Hardie BG, Lee K (2005). "RFM and CLV: Using Iso-Value Curves for Customer Base Analysis." Journal of Marketing Research, 42(4), 415–430.

Fader PS, Hardie BG (2013). "The Gamma-Gamma Model of Monetary Value." URL http://www.brucehardie.com/notes/025/gamma_gamma.pdf.

Examples

# \donttest{

data("apparelTrans")
# Fit pnbd standard model on data, WITH holdout
pnc <- pnbd(clvdata(apparelTrans, time.unit="w",
                    estimation.split=37, date.format="ymd"))

#> Starting estimation...

#> Estimation finished!

# Predict until the end of the holdout period
predict(pnc)

#> Predicting from 2005-09-20 until (incl.) 2006-07-16 (42.86 Weeks).

#>        Id period.first period.last period.length actual.x actual.spending
#>   1:    1   2005-09-20  2006-07-16      42.85714        0            0.00
#>   2:   10   2005-09-20  2006-07-16      42.85714        0            0.00
#>   3:  100   2005-09-20  2006-07-16      42.85714       23         5086.44
#>   4: 1000   2005-09-20  2006-07-16      42.85714       25         4280.24
#>   5: 1001   2005-09-20  2006-07-16      42.85714       11         2914.10
#>  ---                                                                     
#> 246: 1219   2005-09-20  2006-07-16      42.85714       14         3233.78
#> 247:  122   2005-09-20  2006-07-16      42.85714        0            0.00
#> 248: 1220   2005-09-20  2006-07-16      42.85714        0            0.00
#> 249: 1221   2005-09-20  2006-07-16      42.85714        9         1322.94
#> 250: 1222   2005-09-20  2006-07-16      42.85714        0            0.00
#>         PAlive        CET       DERT predicted.Spending predicted.CLV
#>   1: 0.3909486  0.2695175 0.06742966           219.0261      14.76885
#>   2: 0.4683655  1.1455545 0.28660241           208.7025      59.81465
#>   3: 0.9858260 16.2637166 4.06896420           185.2896     753.93666
#>   4: 0.9948489 12.9177435 3.23184652           167.9019     542.63310
#>   5: 0.6492975  5.0094874 1.25330669           234.5086     293.91118
#>  ---                                                                 
#> 246: 0.9859140  4.1431240 1.03655417           204.8505     212.33860
#> 247: 0.3909486  0.2695175 0.06742966           219.0261      14.76885
#> 248: 0.3909486  0.2695175 0.06742966           219.0261      14.76885
#> 249: 0.9762841  4.9600594 1.24094046           206.1343     255.80034
#> 250: 0.4471943  0.7010328 0.17538901           207.6214      36.41452

# Predict until 10 periods (weeks in this case) after
#   the end of the 37 weeks fitting period
predict(pnc, prediction.end = 10) # ends on 2010-11-28

#> Predicting from 2005-09-20 until (incl.) 2005-11-28 (10 Weeks).

#>        Id period.first period.last period.length actual.x actual.spending
#>   1:    1   2005-09-20  2005-11-28            10        0            0.00
#>   2:   10   2005-09-20  2005-11-28            10        0            0.00
#>   3:  100   2005-09-20  2005-11-28            10        6          632.52
#>   4: 1000   2005-09-20  2005-11-28            10        9         1422.00
#>   5: 1001   2005-09-20  2005-11-28            10        3         1029.55
#>  ---                                                                     
#> 246: 1219   2005-09-20  2005-11-28            10        3          216.58
#> 247:  122   2005-09-20  2005-11-28            10        0            0.00
#> 248: 1220   2005-09-20  2005-11-28            10        0            0.00
#> 249: 1221   2005-09-20  2005-11-28            10        4          557.08
#> 250: 1222   2005-09-20  2005-11-28            10        0            0.00
#>         PAlive        CET       DERT predicted.Spending predicted.CLV
#>   1: 0.3909486 0.06948064 0.06742966           219.0261      14.76885
#>   2: 0.4683655 0.29531988 0.28660241           208.7025      59.81465
#>   3: 0.9858260 4.19272818 4.06896420           185.2896     753.93666
#>   4: 0.9948489 3.33014824 3.23184652           167.9019     542.63310
#>   5: 0.6492975 1.29142799 1.25330669           234.5086     293.91118
#>  ---                                                                 
#> 246: 0.9859140 1.06808261 1.03655417           204.8505     212.33860
#> 247: 0.3909486 0.06948064 0.06742966           219.0261      14.76885
#> 248: 0.3909486 0.06948064 0.06742966           219.0261      14.76885
#> 249: 0.9762841 1.27868563 1.24094046           206.1343     255.80034
#> 250: 0.4471943 0.18072375 0.17538901           207.6214      36.41452

# Predict until 31th Dec 2016 with the timepoint as a character
predict(pnc, prediction.end = "2016-12-31")

#> Predicting from 2005-09-20 until (incl.) 2016-12-31 (588.71 Weeks).

#>        Id period.first period.last period.length    PAlive        CET
#>   1:    1   2005-09-20  2016-12-31      588.7143 0.3909486   2.095890
#>   2:   10   2005-09-20  2016-12-31      588.7143 0.4683655   8.908351
#>   3:  100   2005-09-20  2016-12-31      588.7143 0.9858260 126.474033
#>   4: 1000   2005-09-20  2016-12-31      588.7143 0.9948489 100.454230
#>   5: 1001   2005-09-20  2016-12-31      588.7143 0.6492975  38.956045
#>  ---                                                                 
#> 246: 1219   2005-09-20  2016-12-31      588.7143 0.9859140  32.218811
#> 247:  122   2005-09-20  2016-12-31      588.7143 0.3909486   2.095890
#> 248: 1220   2005-09-20  2016-12-31      588.7143 0.3909486   2.095890
#> 249: 1221   2005-09-20  2016-12-31      588.7143 0.9762841  38.571670
#> 250: 1222   2005-09-20  2016-12-31      588.7143 0.4471943   5.451549
#>            DERT predicted.Spending predicted.CLV
#>   1: 0.06742966           219.0261      14.76885
#>   2: 0.28660241           208.7025      59.81465
#>   3: 4.06896420           185.2896     753.93666
#>   4: 3.23184652           167.9019     542.63310
#>   5: 1.25330669           234.5086     293.91118
#>  ---                                            
#> 246: 1.03655417           204.8505     212.33860
#> 247: 0.06742966           219.0261      14.76885
#> 248: 0.06742966           219.0261      14.76885
#> 249: 1.24094046           206.1343     255.80034
#> 250: 0.17538901           207.6214      36.41452

# Predict until 31th Dec 2016 with the timepoint as a Date
predict(pnc, prediction.end = lubridate::ymd("2016-12-31"))

#> Predicting from 2005-09-20 until (incl.) 2016-12-31 (588.71 Weeks).

#>        Id period.first period.last period.length    PAlive        CET
#>   1:    1   2005-09-20  2016-12-31      588.7143 0.3909486   2.095890
#>   2:   10   2005-09-20  2016-12-31      588.7143 0.4683655   8.908351
#>   3:  100   2005-09-20  2016-12-31      588.7143 0.9858260 126.474033
#>   4: 1000   2005-09-20  2016-12-31      588.7143 0.9948489 100.454230
#>   5: 1001   2005-09-20  2016-12-31      588.7143 0.6492975  38.956045
#>  ---                                                                 
#> 246: 1219   2005-09-20  2016-12-31      588.7143 0.9859140  32.218811
#> 247:  122   2005-09-20  2016-12-31      588.7143 0.3909486   2.095890
#> 248: 1220   2005-09-20  2016-12-31      588.7143 0.3909486   2.095890
#> 249: 1221   2005-09-20  2016-12-31      588.7143 0.9762841  38.571670
#> 250: 1222   2005-09-20  2016-12-31      588.7143 0.4471943   5.451549
#>            DERT predicted.Spending predicted.CLV
#>   1: 0.06742966           219.0261      14.76885
#>   2: 0.28660241           208.7025      59.81465
#>   3: 4.06896420           185.2896     753.93666
#>   4: 3.23184652           167.9019     542.63310
#>   5: 1.25330669           234.5086     293.91118
#>  ---                                            
#> 246: 1.03655417           204.8505     212.33860
#> 247: 0.06742966           219.0261      14.76885
#> 248: 0.06742966           219.0261      14.76885
#> 249: 1.24094046           206.1343     255.80034
#> 250: 0.17538901           207.6214      36.41452

# Fit pnbd standard model WITHOUT holdout
pnc <- pnbd(clvdata(apparelTrans, time.unit="w", date.format="ymd"))

#> Starting estimation...

#> Estimation finished!

# This fails, because without holdout, a prediction.end is required
if (FALSE) {
predict(pnc)
}

# Now, predict 10 periods from the end of the last transaction
#   (end of estimation period)
predict(pnc, prediction.end = 10) # ends on 2016-12-17

#> Predicting from 2006-07-16 until (incl.) 2006-09-24 (10 Weeks).

#>        Id period.first period.last period.length     PAlive        CET
#>   1:    1   2006-07-16  2006-09-24            10 0.19161789 0.01673191
#>   2:   10   2006-07-16  2006-09-24            10 0.07059577 0.02243071
#>   3:  100   2006-07-16  2006-09-24            10 0.99543304 4.78886271
#>   4: 1000   2006-07-16  2006-09-24            10 0.99921311 4.57681399
#>   5: 1001   2006-07-16  2006-09-24            10 0.92653166 2.10903146
#>  ---                                                                  
#> 246: 1219   2006-07-16  2006-09-24            10 0.80935629 1.74906523
#> 247:  122   2006-07-16  2006-09-24            10 0.19161789 0.01673191
#> 248: 1220   2006-07-16  2006-09-24            10 0.19161789 0.01673191
#> 249: 1221   2006-07-16  2006-09-24            10 0.88300093 1.50130158
#> 250: 1222   2006-07-16  2006-09-24            10 0.12410715 0.02513502
#>            DERT predicted.Spending predicted.CLV
#>   1: 0.01636545           220.4315      3.607462
#>   2: 0.02193944           216.1576      4.742378
#>   3: 4.68397971           206.0929    965.335142
#>   4: 4.47657517           186.1292    833.221222
#>   5: 2.06284063           233.9810    482.665482
#>  ---                                            
#> 246: 1.71075818           216.8827    371.033786
#> 247: 0.01636545           220.4315      3.607462
#> 248: 0.01636545           220.4315      3.607462
#> 249: 1.46842091           204.4750    300.255303
#> 250: 0.02458453           215.8908      5.307575

# }