Predict CLV from a fitted transaction model

Probabilistic customer attrition models predict in general three expected characteristics for every customer:

"conditional expected transactions" (CET), which is the number of transactions to expect from a customer during the prediction period,

"probability of a customer being alive" (PAlive) at the end of the estimation period and

"discounted expected residual transactions" (DERT) for every customer, which is the total number of transactions for the residual lifetime of a customer discounted to the end of the estimation period. In the case of time-varying covariates, instead of DERT, "discounted expected conditional transactions" (DECT) is predicted. DECT does only cover a finite time horizon in contrast to DERT. For continuous.discount.factor=0, DECT corresponds to CET.

In order to derive a monetary value such as CLV, customer spending has to be considered. If the clv.data object contains spending information, customer spending can be predicted using a Gamma/Gamma spending model for parameter predict.spending and the predicted CLV is be calculated (if the transaction model supports DERT/DECT). In this case, the prediction additionally contains the following two columns:

"predicted.mean.spending", the mean spending per transactions as predicted by the spending model.

"CLV", the customer lifetime value. CLV is the product of DERT/DECT and predicted spending.

Uncertainty estimates are available for all predicted quantities using bootstrapping.

New customer prediction

The fitted model can also be used to predict the number of transactions a single, average newly alive customer is expected to make at the moment of the first transaction ("coming alive"). For covariate models, the prediction is for an average customer with the given covariates.

The individual-level unconditional expectation that is also used for the tracking plot is used to obtain this prediction. For models without covariates, the prediction hence is the same for all customers and independent of when a customer comes alive. For models with covariates, the prediction is the same for all customers with the same covariates.

The data on which the model was fit and which is stored in it is NOT used for this prediction. See examples and newcustomer for more details.

# S3 method for class 'clv.fitted.transactions'
predict(
  object,
  newdata = NULL,
  prediction.end = NULL,
  predict.spending = gg,
  continuous.discount.factor = log(1 + 0.1),
  uncertainty = c("none", "boots"),
  level = 0.9,
  num.boots = 100,
  verbose = TRUE,
  ...
)

# S4 method for class 'clv.fitted.transactions'
predict(
  object,
  newdata = NULL,
  prediction.end = NULL,
  predict.spending = gg,
  continuous.discount.factor = log(1 + 0.1),
  uncertainty = c("none", "boots"),
  level = 0.9,
  num.boots = 100,
  verbose = TRUE,
  ...
)

Arguments

object

A fitted clv transaction model for which prediction is desired.

newdata

A clv data object or data for the new customer prediction (see newcustomer) for which predictions should be made with the fitted model. If none or NULL is given, predictions are made for the data on which the model was fit.

prediction.end

Until what point in time to predict. This can be the number of periods (numeric) or a form of date/time object. See details.

predict.spending

Whether and how to predict spending and based on it also CLV, if possible. See details.

continuous.discount.factor

continuous discount factor to use to calculate DERT/DECT. Defaults to a 10% continuous annual rate. See details.

uncertainty

Method to produce confidence intervals of the predictions (parameter uncertainty). Either "none" (default) or "boots".

level

Required confidence level, if uncertainty="boots".

num.boots

Number of bootstrap repetitions, if uncertainty="boots". A low number may not produce intervals for all customers if they are not sampled.

verbose

Show details about the running of the function.

...

Ignored

Value

An object of class data.table with columns:

Id

The respective customer identifier

period.first

First timepoint of prediction period

period.last

Last timepoint of prediction period

period.length

Number of time units covered by the period indicated by period.first and period.last (including both ends).

PAlive

Probability to be alive at the end of the estimation period

CET

The Conditional Expected Transactions: The number of transactions expected until prediction.end.

DERT or DECT

Discounted Expected Residual Transactions or Discounted Expected Conditional Transactions for dynamic covariates models

actual.x

Actual number of transactions until prediction.end. Only if there is a holdout period and the prediction ends in it, otherwise not reported.

actual.total.spending

Actual total spending until prediction.end. Only if there is a holdout period and the prediction ends in it, otherwise not reported.

predicted.mean.spending

The mean spending per transactions as predicted by the spending model.

predicted.total.spending

The predicted total spending until prediction.end (CET*predicted.mean.spending).

predicted.CLV

Customer Lifetime Value based on DERT/DECT and predicted.mean.spending.

If predicting for new customers (using newcustomer()), a numeric scalar indicating the expected number of transactions is returned instead.

Details

predict.spending indicates whether to predict customers' spending and if so, the spending model to use. Accepted inputs are either a logical (TRUE/FALSE), a method to fit a spending model (i.e. gg), or an already fitted spending model. If provided TRUE, a Gamma-Gamma model is fit with default options. If argument newdata is provided, the spending model is fit on newdata. Predicting spending is only possible if the transaction data contains spending information. See examples for illustrations of valid inputs.

The newdata argument has to be a clv data object of the exact same class as the data object on which the model was fit. In case the model was fit with covariates, newdata needs to contain identically named covariate data.

The use case for newdata is mainly two-fold: First, to estimate model parameters only on a sample of the data and then use the fitted model object to predict or plot for the full data set provided through newdata. Second, for models with dynamic covariates, to provide a clv data object with longer covariates than contained in the data on which the model was estimated what allows to predict or plot further. When providing newdata, some models might require additional steps that can significantly increase runtime.

To predict for new customers, the output of newcustomer has to be given to newdata. See examples.

prediction.end indicates until when to predict or plot and can be given as either a point in time (of class Date, POSIXct, or character) or the number of periods. If prediction.end is of class character, the date/time format set when creating the data object is used for parsing. If prediction.end is the number of periods, the end of the fitting period serves as the reference point from which periods are counted. Only full periods may be specified. If prediction.end is omitted or NULL, it defaults to the end of the holdout period if present and to the end of the estimation period otherwise.

The first prediction period is defined to start right after the end of the estimation period. If for example weekly time units are used and the estimation period ends on Sunday 2019-01-01, then the first day of the first prediction period is Monday 2019-01-02. Each prediction period includes a total of 7 days and the first prediction period therefore will end on, and include, Sunday 2019-01-08. Subsequent prediction periods again start on Mondays and end on Sundays. If prediction.end indicates a timepoint on which to end, this timepoint is included in the prediction period.

continuous.discount.factor is the continuous rate used to discount the expected residual transactions (DERT/DECT). An annual rate of (100 x d)% equals a continuous rate delta = ln(1+d). To account for time units which are not annual, the continuous rate has to be further adjusted to delta=ln(1+d)/k, where k are the number of time units in a year.

Uncertainty Estimates

Bootstrapping is used to provide confidence intervals of all predicted metrics. These provide an estimate of parameter uncertainty. To create bootstrapped data, customer ids are sampled with replacement until reaching original length and all transactions of the sampled customers are used to create a new clv.data object. A new model is fit on the bootstrapped data with the same specification as object (incl. start parameters and `optimx.args`) and it is then used to predict on this data. All prediction parameters, incl prediction.end and continuous.discount.factor, are forwarded to the prediction on the bootstrapped data. Per customer, confidence intervals of each predicted metric are created using a "reversed quantile" approach. See clv.bootstrapped.apply to create a custom bootstrapping procedure.

See also

Examples

# \donttest{

data("apparelTrans")
# Fit pnbd standard model on data, WITH holdout
apparel.holdout <- clvdata(apparelTrans, time.unit="w",
                           estimation.split=37, date.format="ymd")
apparel.pnbd <- pnbd(apparel.holdout)
#> Starting estimation...
#> Estimation finished!

# Predict until the end of the holdout period
predict(apparel.pnbd)
#> Predicting from 2005-09-20 until (incl.) 2006-07-16 (42.86 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length actual.x
#>      <char>       <Date>      <Date>         <num>    <int>
#>   1:      1   2005-09-20  2006-07-16      42.85714        0
#>   2:     10   2005-09-20  2006-07-16      42.85714        0
#>   3:    100   2005-09-20  2006-07-16      42.85714       23
#>   4:   1000   2005-09-20  2006-07-16      42.85714       25
#>   5:   1001   2005-09-20  2006-07-16      42.85714       11
#>  ---                                                       
#> 246:   1219   2005-09-20  2006-07-16      42.85714       14
#> 247:    122   2005-09-20  2006-07-16      42.85714        0
#> 248:   1220   2005-09-20  2006-07-16      42.85714        0
#> 249:   1221   2005-09-20  2006-07-16      42.85714        9
#> 250:   1222   2005-09-20  2006-07-16      42.85714        0
#>      actual.total.spending    PAlive        CET      DERT
#>                      <num>     <num>      <num>     <num>
#>   1:                  0.00 0.3908705  0.2695076 0.0705440
#>   2:                  0.00 0.4683273  1.1454855 0.2998325
#>   3:                737.53 0.9858250 16.2632775 4.2569363
#>   4:               1120.41 0.9948485 12.9174313 3.3811563
#>   5:                364.00 0.6492819  5.0092842 1.3111874
#>  ---                                                     
#> 246:                413.76 0.9859129  4.1431118 1.0844655
#> 247:                  0.00 0.3908705  0.2695076 0.0705440
#> 248:                  0.00 0.3908705  0.2695076 0.0705440
#> 249:                302.65 0.9762824  4.9600134 1.2982907
#> 250:                  0.00 0.4471444  0.7009915 0.1834855
#>      predicted.mean.spending predicted.total.spending predicted.CLV
#>                        <num>                    <num>         <num>
#>   1:                40.39473                 10.88669      2.849606
#>   2:                57.49121                 65.85534     17.237731
#>   3:                43.83628                712.92159    186.608253
#>   4:                43.46097                561.40408    146.948327
#>   5:                46.17637                231.31054     60.545866
#>  ---                                                               
#> 246:                33.20726                137.58140     36.012131
#> 247:                40.39473                 10.88669      2.849606
#> 248:                40.39473                 10.88669      2.849606
#> 249:                34.00868                168.68352     44.153154
#> 250:                48.80122                 34.20924      8.954317

# Predict until 10 periods (weeks in this case) after
#   the end of the 37 weeks fitting period
predict(apparel.pnbd, prediction.end = 10) # ends on 2010-11-28
#> Predicting from 2005-09-20 until (incl.) 2005-11-28 (10 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length actual.x
#>      <char>       <Date>      <Date>         <int>    <int>
#>   1:      1   2005-09-20  2005-11-28            10        0
#>   2:     10   2005-09-20  2005-11-28            10        0
#>   3:    100   2005-09-20  2005-11-28            10        6
#>   4:   1000   2005-09-20  2005-11-28            10        9
#>   5:   1001   2005-09-20  2005-11-28            10        3
#>  ---                                                       
#> 246:   1219   2005-09-20  2005-11-28            10        3
#> 247:    122   2005-09-20  2005-11-28            10        0
#> 248:   1220   2005-09-20  2005-11-28            10        0
#> 249:   1221   2005-09-20  2005-11-28            10        4
#> 250:   1222   2005-09-20  2005-11-28            10        0
#>      actual.total.spending    PAlive        CET      DERT
#>                      <num>     <num>      <num>     <num>
#>   1:                  0.00 0.3908705 0.06947848 0.0705440
#>   2:                  0.00 0.4683273 0.29530372 0.2998325
#>   3:                134.15 0.9858250 4.19263832 4.2569363
#>   4:                318.20 0.9948485 3.33008629 3.3811563
#>   5:                118.22 0.6492819 1.29138281 1.3111874
#>  ---                                                     
#> 246:                 93.92 0.9859129 1.06808542 1.0844655
#> 247:                  0.00 0.3908705 0.06947848 0.0705440
#> 248:                  0.00 0.3908705 0.06947848 0.0705440
#> 249:                 91.80 0.9762824 1.27868090 1.2982907
#> 250:                  0.00 0.4471444 0.18071411 0.1834855
#>      predicted.mean.spending predicted.total.spending predicted.CLV
#>                        <num>                    <num>         <num>
#>   1:                40.39473                 2.806565      2.849606
#>   2:                57.49121                16.977367     17.237731
#>   3:                43.83628               183.789668    186.608253
#>   4:                43.46097               144.728777    146.948327
#>   5:                46.17637                59.631364     60.545866
#>  ---                                                               
#> 246:                33.20726                35.468193     36.012131
#> 247:                40.39473                 2.806565      2.849606
#> 248:                40.39473                 2.806565      2.849606
#> 249:                34.00868                43.486252     44.153154
#> 250:                48.80122                 8.819068      8.954317

# Predict until 31th Dec 2016 with the timepoint as a character
predict(apparel.pnbd, prediction.end = "2016-12-31")
#> Predicting from 2005-09-20 until (incl.) 2016-12-31 (588.71 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length    PAlive        CET
#>      <char>       <Date>      <Date>         <num>     <num>      <num>
#>   1:      1   2005-09-20  2016-12-31      588.7143 0.3908705   2.095780
#>   2:     10   2005-09-20  2016-12-31      588.7143 0.4683273   8.907673
#>   3:    100   2005-09-20  2016-12-31      588.7143 0.9858250 126.468606
#>   4:   1000   2005-09-20  2016-12-31      588.7143 0.9948485 100.450204
#>   5:   1001   2005-09-20  2016-12-31      588.7143 0.6492819  38.953845
#>  ---                                                                   
#> 246:   1219   2005-09-20  2016-12-31      588.7143 0.9859129  32.218203
#> 247:    122   2005-09-20  2016-12-31      588.7143 0.3908705   2.095780
#> 248:   1220   2005-09-20  2016-12-31      588.7143 0.3908705   2.095780
#> 249:   1221   2005-09-20  2016-12-31      588.7143 0.9762824  38.570699
#> 250:   1222   2005-09-20  2016-12-31      588.7143 0.4471444   5.451141
#>           DERT predicted.mean.spending predicted.total.spending predicted.CLV
#>          <num>                   <num>                    <num>         <num>
#>   1: 0.0705440                40.39473                 84.65847      2.849606
#>   2: 0.2998325                57.49121                512.11286     17.237731
#>   3: 4.2569363                43.83628               5543.91326    186.608253
#>   4: 3.3811563                43.46097               4365.66317    146.948327
#>   5: 1.3111874                46.17637               1798.74698     60.545866
#>  ---                                                                         
#> 246: 1.0844655                33.20726               1069.87835     36.012131
#> 247: 0.0705440                40.39473                 84.65847      2.849606
#> 248: 0.0705440                40.39473                 84.65847      2.849606
#> 249: 1.2982907                34.00868               1311.73864     44.153154
#> 250: 0.1834855                48.80122                266.02230      8.954317

# Predict until 31th Dec 2016 with the timepoint as a Date
predict(apparel.pnbd, prediction.end = lubridate::ymd("2016-12-31"))
#> Predicting from 2005-09-20 until (incl.) 2016-12-31 (588.71 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length    PAlive        CET
#>      <char>       <Date>      <Date>         <num>     <num>      <num>
#>   1:      1   2005-09-20  2016-12-31      588.7143 0.3908705   2.095780
#>   2:     10   2005-09-20  2016-12-31      588.7143 0.4683273   8.907673
#>   3:    100   2005-09-20  2016-12-31      588.7143 0.9858250 126.468606
#>   4:   1000   2005-09-20  2016-12-31      588.7143 0.9948485 100.450204
#>   5:   1001   2005-09-20  2016-12-31      588.7143 0.6492819  38.953845
#>  ---                                                                   
#> 246:   1219   2005-09-20  2016-12-31      588.7143 0.9859129  32.218203
#> 247:    122   2005-09-20  2016-12-31      588.7143 0.3908705   2.095780
#> 248:   1220   2005-09-20  2016-12-31      588.7143 0.3908705   2.095780
#> 249:   1221   2005-09-20  2016-12-31      588.7143 0.9762824  38.570699
#> 250:   1222   2005-09-20  2016-12-31      588.7143 0.4471444   5.451141
#>           DERT predicted.mean.spending predicted.total.spending predicted.CLV
#>          <num>                   <num>                    <num>         <num>
#>   1: 0.0705440                40.39473                 84.65847      2.849606
#>   2: 0.2998325                57.49121                512.11286     17.237731
#>   3: 4.2569363                43.83628               5543.91326    186.608253
#>   4: 3.3811563                43.46097               4365.66317    146.948327
#>   5: 1.3111874                46.17637               1798.74698     60.545866
#>  ---                                                                         
#> 246: 1.0844655                33.20726               1069.87835     36.012131
#> 247: 0.0705440                40.39473                 84.65847      2.849606
#> 248: 0.0705440                40.39473                 84.65847      2.849606
#> 249: 1.2982907                34.00868               1311.73864     44.153154
#> 250: 0.1834855                48.80122                266.02230      8.954317

# Predict future transactions but not spending and CLV
predict(apparel.pnbd, predict.spending = FALSE)
#> Predicting from 2005-09-20 until (incl.) 2006-07-16 (42.86 Weeks).
#> Key: <Id>
#>          Id period.first period.last period.length actual.x
#>      <char>       <Date>      <Date>         <num>    <int>
#>   1:      1   2005-09-20  2006-07-16      42.85714        0
#>   2:     10   2005-09-20  2006-07-16      42.85714        0
#>   3:    100   2005-09-20  2006-07-16      42.85714       23
#>   4:   1000   2005-09-20  2006-07-16      42.85714       25
#>   5:   1001   2005-09-20  2006-07-16      42.85714       11
#>  ---                                                       
#> 246:   1219   2005-09-20  2006-07-16      42.85714       14
#> 247:    122   2005-09-20  2006-07-16      42.85714        0
#> 248:   1220   2005-09-20  2006-07-16      42.85714        0
#> 249:   1221   2005-09-20  2006-07-16      42.85714        9
#> 250:   1222   2005-09-20  2006-07-16      42.85714        0
#>      actual.total.spending    PAlive        CET      DERT
#>                      <num>     <num>      <num>     <num>
#>   1:                  0.00 0.3908705  0.2695076 0.0705440
#>   2:                  0.00 0.4683273  1.1454855 0.2998325
#>   3:                737.53 0.9858250 16.2632775 4.2569363
#>   4:               1120.41 0.9948485 12.9174313 3.3811563
#>   5:                364.00 0.6492819  5.0092842 1.3111874
#>  ---                                                     
#> 246:                413.76 0.9859129  4.1431118 1.0844655
#> 247:                  0.00 0.3908705  0.2695076 0.0705440
#> 248:                  0.00 0.3908705  0.2695076 0.0705440
#> 249:                302.65 0.9762824  4.9600134 1.2982907
#> 250:                  0.00 0.4471444  0.7009915 0.1834855

# Predict spending by fitting a Gamma-Gamma model
predict(apparel.pnbd, predict.spending = gg)
#> Predicting from 2005-09-20 until (incl.) 2006-07-16 (42.86 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length actual.x
#>      <char>       <Date>      <Date>         <num>    <int>
#>   1:      1   2005-09-20  2006-07-16      42.85714        0
#>   2:     10   2005-09-20  2006-07-16      42.85714        0
#>   3:    100   2005-09-20  2006-07-16      42.85714       23
#>   4:   1000   2005-09-20  2006-07-16      42.85714       25
#>   5:   1001   2005-09-20  2006-07-16      42.85714       11
#>  ---                                                       
#> 246:   1219   2005-09-20  2006-07-16      42.85714       14
#> 247:    122   2005-09-20  2006-07-16      42.85714        0
#> 248:   1220   2005-09-20  2006-07-16      42.85714        0
#> 249:   1221   2005-09-20  2006-07-16      42.85714        9
#> 250:   1222   2005-09-20  2006-07-16      42.85714        0
#>      actual.total.spending    PAlive        CET      DERT
#>                      <num>     <num>      <num>     <num>
#>   1:                  0.00 0.3908705  0.2695076 0.0705440
#>   2:                  0.00 0.4683273  1.1454855 0.2998325
#>   3:                737.53 0.9858250 16.2632775 4.2569363
#>   4:               1120.41 0.9948485 12.9174313 3.3811563
#>   5:                364.00 0.6492819  5.0092842 1.3111874
#>  ---                                                     
#> 246:                413.76 0.9859129  4.1431118 1.0844655
#> 247:                  0.00 0.3908705  0.2695076 0.0705440
#> 248:                  0.00 0.3908705  0.2695076 0.0705440
#> 249:                302.65 0.9762824  4.9600134 1.2982907
#> 250:                  0.00 0.4471444  0.7009915 0.1834855
#>      predicted.mean.spending predicted.total.spending predicted.CLV
#>                        <num>                    <num>         <num>
#>   1:                40.39473                 10.88669      2.849606
#>   2:                57.49121                 65.85534     17.237731
#>   3:                43.83628                712.92159    186.608253
#>   4:                43.46097                561.40408    146.948327
#>   5:                46.17637                231.31054     60.545866
#>  ---                                                               
#> 246:                33.20726                137.58140     36.012131
#> 247:                40.39473                 10.88669      2.849606
#> 248:                40.39473                 10.88669      2.849606
#> 249:                34.00868                168.68352     44.153154
#> 250:                48.80122                 34.20924      8.954317

# Fit a spending model separately and use it to predict spending
apparel.gg <- gg(apparel.holdout, remove.first.transaction = FALSE)
#> Starting estimation...
#> Estimation finished!
predict(apparel.pnbd, predict.spending = apparel.gg)
#> Predicting from 2005-09-20 until (incl.) 2006-07-16 (42.86 Weeks).
#> Key: <Id>
#>          Id period.first period.last period.length actual.x
#>      <char>       <Date>      <Date>         <num>    <int>
#>   1:      1   2005-09-20  2006-07-16      42.85714        0
#>   2:     10   2005-09-20  2006-07-16      42.85714        0
#>   3:    100   2005-09-20  2006-07-16      42.85714       23
#>   4:   1000   2005-09-20  2006-07-16      42.85714       25
#>   5:   1001   2005-09-20  2006-07-16      42.85714       11
#>  ---                                                       
#> 246:   1219   2005-09-20  2006-07-16      42.85714       14
#> 247:    122   2005-09-20  2006-07-16      42.85714        0
#> 248:   1220   2005-09-20  2006-07-16      42.85714        0
#> 249:   1221   2005-09-20  2006-07-16      42.85714        9
#> 250:   1222   2005-09-20  2006-07-16      42.85714        0
#>      actual.total.spending    PAlive        CET      DERT
#>                      <num>     <num>      <num>     <num>
#>   1:                  0.00 0.3908705  0.2695076 0.0705440
#>   2:                  0.00 0.4683273  1.1454855 0.2998325
#>   3:                737.53 0.9858250 16.2632775 4.2569363
#>   4:               1120.41 0.9948485 12.9174313 3.3811563
#>   5:                364.00 0.6492819  5.0092842 1.3111874
#>  ---                                                     
#> 246:                413.76 0.9859129  4.1431118 1.0844655
#> 247:                  0.00 0.3908705  0.2695076 0.0705440
#> 248:                  0.00 0.3908705  0.2695076 0.0705440
#> 249:                302.65 0.9762824  4.9600134 1.2982907
#> 250:                  0.00 0.4471444  0.7009915 0.1834855
#>      predicted.mean.spending predicted.total.spending predicted.CLV
#>                        <num>                    <num>         <num>
#>   1:                70.27718                18.940234      4.957634
#>   2:                61.78767                70.776884     18.525951
#>   3:                42.33307               688.474460    180.209182
#>   4:                41.93464               541.687877    141.787583
#>   5:                45.28636               226.852233     59.378899
#>  ---                                                               
#> 246:                29.90681               123.907253     32.432902
#> 247:                34.95251                 9.419966      2.465690
#> 248:                37.82832                10.195020      2.668561
#> 249:                30.88611               153.195497     40.099142
#> 250:                48.29395                33.853644      8.861240

# Fit pnbd standard model WITHOUT holdout
pnc <- pnbd(clvdata(apparelTrans, time.unit="w", date.format="ymd"))
#> Starting estimation...
#> Estimation finished!

# This fails, because without holdout, a prediction.end is required
if (FALSE) { # \dontrun{
predict(pnc)
} # }

# But it works if providing a prediction.end
predict(pnc, prediction.end = 10) # ends on 2016-12-17
#> Predicting from 2006-07-17 until (incl.) 2006-09-24 (10 Weeks).
#> Estimating gg model to predict spending...
#> Starting estimation...
#> Estimation finished!
#> Key: <Id>
#>          Id period.first period.last period.length     PAlive        CET
#>      <char>       <Date>      <Date>         <int>      <num>      <num>
#>   1:      1   2006-07-17  2006-09-24            10 0.19162799 0.01673257
#>   2:     10   2006-07-17  2006-09-24            10 0.07060463 0.02243351
#>   3:    100   2006-07-17  2006-09-24            10 0.99543387 4.78888430
#>   4:   1000   2006-07-17  2006-09-24            10 0.99921326 4.57683140
#>   5:   1001   2006-07-17  2006-09-24            10 0.92654420 2.10906702
#>  ---                                                                    
#> 246:   1219   2006-07-17  2006-09-24            10 0.80938488 1.74913277
#> 247:    122   2006-07-17  2006-09-24            10 0.19162799 0.01673257
#> 248:   1220   2006-07-17  2006-09-24            10 0.19162799 0.01673257
#> 249:   1221   2006-07-17  2006-09-24            10 0.88302000 1.50133868
#> 250:   1222   2006-07-17  2006-09-24            10 0.12411992 0.02513753
#>            DERT predicted.mean.spending predicted.total.spending predicted.CLV
#>           <num>                   <num>                    <num>         <num>
#>   1: 0.01713505                38.64708                 0.646665     0.6622198
#>   2: 0.02297312                50.23072                 1.126851     1.1539564
#>   3: 4.90407538                37.91699               181.580090   185.9477891
#>   4: 4.68692179                43.53497               199.252218   204.0449995
#>   5: 2.15979819                39.80794                83.957620    85.9771239
#>  ---                                                                          
#> 246: 1.79120614                31.80966                55.639319    56.9776588
#> 247: 0.01713505                38.64708                 0.646665     0.6622198
#> 248: 0.01713505                38.64708                 0.646665     0.6622198
#> 249: 1.53745165                34.22209                51.378944    52.6148051
#> 250: 0.02574218                44.10134                 1.108599     1.1352646

# Predict num transactions for a newly alive customer
# in the next 3.45 weeks
# See ?newcustomer() for more examples
predict(apparel.pnbd, newdata = newcustomer(num.periods=3.45))
#> [1] 0.4707139

# }