Summarizing a fitted CLV model

Summary method for fitted CLV models that provides statistics about the estimated parameters and information about the optimization process. If multiple optimization methods were used (for example if specified in parameter optimx.args), all information here refers to the last method/row of the resulting optimx object.

Note that for the main model coefficients (all coefs not for covariates), the significance indicators z-val and p-values are set to NA because they are by definition always strictly positive and hypothesis test relative to a null of 0 does not make sense.

# S3 method for class 'clv.fitted'
summary(object, ...)

# S3 method for class 'clv.fitted.transactions.static.cov'
summary(object, ...)

# S3 method for class 'summary.clv.fitted'
print(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

Arguments

object

A fitted CLV model

...

Ignored for summary, forwarded to printCoefmat for print.

x

an object of class "summary.clv.no.covariates", usually, a result of a call to summary.clv.no.covariates.

digits

the number of significant digits to use when printing.

signif.stars

logical. If TRUE, ‘significance stars’ are printed for each coefficient.

Value

This function computes and returns a list of summary information of the fitted model given in object. It returns a list of class summary.clv.no.covariates that contains the following components:

name.model

the name of the fitted model.

call

The call used to fit the model.

tp.estimation.start

Date or POSIXct indicating when the fitting period started.

tp.estimation.end

Date or POSIXct indicating when the fitting period ended.

estimation.period.in.tu

Length of fitting period in time.units.

time.unit

Time unit that defines a single period.

coefficients

a px4 matrix with columns for the estimated coefficients, its standard error, the t-statistic and corresponding (two-sided) p-value.

estimated.LL

the value of the log-likelihood function at the found solution.

AIC

Akaike's An Information Criterion for the fitted model.

BIC

Schwarz' Bayesian Information Criterion for the fitted model.

KKT1

Karush-Kuhn-Tucker optimality conditions of the first order, as returned by optimx.

KKT2

Karush-Kuhn-Tucker optimality conditions of the second order, as returned by optimx.

fevals

The number of calls to the log-likelihood function during optimization.

method

The last method used to obtain the final solution.

additional.options

A list of additional options used for model fitting.

Correlation

Whether the correlation between the purchase and the attrition process was estimated.

estimated.param.cor

Correlation coefficient measuring the correlation between the two processes, if used.

For models fits with static covariates, the list additionally is of class summary.clv.static.covariates and the list in additional.options contains the following elements:

additional.options

Regularization

Whether L2 regularization for parameters of contextual factors was used.

lambda.life

The regularization lambda used for the parameters of the Lifetime process, if used.

lambda.trans

The regularization lambda used for the parameters of the Transaction process, if used.

Constraint covs

Whether any covariate parameters were forced to be the same for both processes.

Constraint params

Name of the covariate parameters which were constraint, if used.

See also

Examples

# \donttest{

data("apparelTrans")

# Fit pnbd standard model, no covariates
clv.data.apparel <- clvdata(apparelTrans, time.unit="w",
                                estimation.split=52, date.format="ymd")
pnbd.apparel <- pnbd(clv.data.apparel)
#> Starting estimation...
#> Estimation finished!

# summary about model fit
summary(pnbd.apparel)
#> Pareto/NBD Standard  Model 
#> 
#> Call:
#> pnbd(clv.data = clv.data.apparel)
#> 
#> Fitting period:                               
#> Estimation start  2005-01-02   
#> Estimation end    2006-01-01   
#> Estimation length 52.0000 Weeks
#> 
#> Coefficients:
#>       Estimate Std. Error z-val Pr(>|z|)
#> r       1.3080     0.2127    NA       NA
#> alpha  46.3352     7.5456    NA       NA
#> s       0.4794     1.0791    NA       NA
#> beta   70.2292   202.6675    NA       NA
#> 
#> Optimization info:                 
#> LL     -3481.9782
#> AIC    6971.9564 
#> BIC    6989.5442 
#> KKT 1  TRUE      
#> KKT 2  TRUE      
#> fevals 25.0000   
#> Method L-BFGS-B  
#> 
#> Used Options:                 
#> Correlation FALSE

# Add static covariate data
data("apparelStaticCov")
data.apparel.cov  <-
   SetStaticCovariates(clv.data.apparel,
                       data.cov.life  = apparelStaticCov,
                       names.cov.life = "Gender",
                       data.cov.trans = apparelStaticCov,
                       names.cov.trans = "Gender",
                       name.id = "Id")

# fit model with covariates and regualization
pnbd.apparel.cov <- pnbd(data.apparel.cov,
                         reg.lambdas = c(life=2, trans=4))
#> Starting estimation...
#> Estimation finished!

# additional summary about covariate parameters
#   and used regularization
summary(pnbd.apparel.cov)
#> Pareto/NBD with Static Covariates  Model 
#> 
#> Call:
#> pnbd(clv.data = data.apparel.cov, reg.lambdas = c(life = 2, trans = 4))
#> 
#> Fitting period:                               
#> Estimation start  2005-01-02   
#> Estimation end    2006-01-01   
#> Estimation length 52.0000 Weeks
#> 
#> Coefficients:
#>                Estimate Std. Error  z-val Pr(>|z|)
#> r             1.313e+00  5.238e+00     NA       NA
#> alpha         4.669e+01  1.871e+02     NA       NA
#> s             4.808e-01  2.655e+01     NA       NA
#> beta          7.029e+01  4.974e+03     NA       NA
#> life.Gender  -1.900e-03  4.975e-01 -0.004    0.997
#> trans.Gender  7.675e-03  4.975e-01  0.015    0.988
#> 
#> Optimization info:               
#> LL     -5.8031 
#> AIC    23.6061 
#> BIC    49.9877 
#> KKT 1  TRUE    
#> KKT 2  TRUE    
#> fevals 78.0000 
#> Method L-BFGS-B
#> 
#> Used Options:                      
#> Correlation     FALSE 
#> Regularization  TRUE  
#>    lambda.life  2.0000
#>    lambda.trans 4.0000
#> Constraint covs FALSE 
# }