Calculate Variance-Covariance Matrix for CLV Models fitted with Maximum Likelihood Estimation

Returns the variance-covariance matrix of the parameters of the fitted model object. The variance-covariance matrix is derived from the Hessian that results from the optimization procedure. First, the Moore–Penrose generalized inverse of the Hessian is used to obtain an estimate of the variance-covariance matrix. Next, because some parameters may be transformed for the purpose of restricting their value during the log-likelihood estimation, the variance estimates are adapted to be comparable to the reported coefficient estimates. If the result is not positive definite, nearPD is used with standard settings to find the nearest positive definite matrix.

If multiple estimation methods were used, the Hessian of the last method is used.

# S3 method for clv.fitted
vcov(object, ...)

Arguments

object
a fitted clv model object

...
Ignored

Value

A matrix of the estimated covariances between the parameters of the model. The row and column names correspond to the parameter names given by the coef method.

See also