Predict customers' future spending

Predict customer's future mean spending per transaction and compare it to the actual mean spending in the holdout period.

# S3 method for class 'clv.fitted.spending'
predict(
  object,
  newdata = NULL,
  uncertainty = c("none", "boots"),
  level = 0.9,
  num.boots = 100,
  verbose = TRUE,
  ...
)

# S4 method for class 'clv.fitted.spending'
predict(
  object,
  newdata = NULL,
  uncertainty = c("none", "boots"),
  level = 0.9,
  num.boots = 100,
  verbose = TRUE,
  ...
)

Arguments

object

A fitted spending model for which prediction is desired.

newdata

A clv data object for which predictions should be made with the fitted model. If none or NULL is given, predictions are made for the data on which the model was fit.

uncertainty

Method to produce confidence intervals of the predictions (parameter uncertainty). Either "none" (default) or "boots".

level

Required confidence level, if uncertainty="boots".

num.boots

Number of bootstrap repetitions, if uncertainty="boots". A low number may not produce intervals for all customers if they are not sampled.

verbose

Show details about the running of the function.

...

Ignored

Value

An object of class data.table with columns:

Id

The respective customer identifier

actual.mean.spending

Actual mean spending per transaction in the holdout period. Only if there is a holdout period otherwise it is not reported.

predicted.mean.spending

The mean spending per transaction as predicted by the fitted spending model.

Details

If newdata is provided, the individual customer statistics underlying the model are calculated the same way as when the model was fit initially. Hence, if remove.first.transaction was TRUE, this will be applied to newdata as well.

See also

Examples

# \donttest{
data("apparelTrans")

# Fit gg model on data
apparel.holdout <- clvdata(apparelTrans, time.unit="w",
                           estimation.split = 52, date.format = "ymd")
apparel.gg <- gg(apparel.holdout)
#> Starting estimation...
#> Estimation finished!

# Predict customers' future mean spending per transaction
predict(apparel.gg)
#> Key: <Id>
#>          Id actual.mean.spending predicted.mean.spending
#>      <char>                <num>                   <num>
#>   1:      1            104.82000                63.77959
#>   2:     10             31.65500                38.10316
#>   3:    100             37.03889                37.12434
#>   4:    101              0.00000                33.08404
#>   5:    102              0.00000                37.12434
#>  ---                                                    
#> 596:     95             22.76000                28.29614
#> 597:     96             84.53667                37.12434
#> 598:     97              0.00000                37.12434
#> 599:     98              0.00000                34.87739
#> 600:     99             13.99000                16.11315

# }