Summarizing a CLV data object

Summary method for objects of class clv.data that provides information about the estimation and possible holdout sample, and descriptive statistics of the transaction data.

# S3 method for clv.data
summary(object, ...)

# S3 method for summary.clv.data
print(x, digits = max(3L, getOption("digits") - 3L), ...)

# S3 method for clv.data.dynamic.covariates
summary(object, ...)

# S3 method for summary.clv.data.dynamic.covariates
print(x, digits = max(3L, getOption("digits") - 3L), ...)

# S3 method for clv.data.static.covariates
summary(object, ...)

# S3 method for summary.clv.data.static.covariates
print(x, digits = max(3L, getOption("digits") - 3L), ...)

Arguments

object
A CLV data object containing transactional data and potentially also contextual factors.

...
Ignored.

x
An object of class "summary.clv.data", usually, a result of a call to summary.clv.data.

digits
The number of significant digits to use when printing.

Value

This function computes and returns summary statistics of the transactional and covariates data given in object. This is a list of class summary.clv.data and contains the elements:

name

Human readable description of the type of data.

summary.clv.tim

Summary information about the stored clv.time object.

descriptives.transactions

A data.table with summary statistics of the transactions overall and in the estimation and holdout sample.

For static covariates data, the list additionally is of class summary.clv.data.static.covariates and further contains the elements:
names.cov.data.trans

Names of the covariates for the Transaction process.

names.cov.data.life

Names of the covariates for the Lifetime process.

See also

Examples

# \donttest{

data("apparelTrans")
clv.data.apparel <- clvdata(apparelTrans, date.format = "ymd",
                            time.unit = "w",
                            estimation.split = 40)

# summary of transaction data and split
summary(clv.data.apparel)

#> CLV Transaction Data 
#>                                 
#> Time unit         Weeks         
#> Estimation length 40.0000 Weeks 
#> Holdout length    39.71429 Weeks
#> 
#> Transaction Data Summary 
#>                                    Estimation      Holdout         Total     
#> Number of customers                -               -               250       
#> First Transaction in period        2005-01-03      2005-10-11      2005-01-03
#> Last Transaction in period         2005-10-10      2006-07-16      2006-07-16
#> Total # Transactions               1311            946             2257      
#> Mean # Transactions per cust       5.244           8.226           9.028     
#> (SD)                               6.082           8.934           12.603    
#> Mean Spending per Transaction      39.436          38.519          39.051    
#> (SD)                               42.649          59.723          50.503    
#> Total Spending                     51700.490       36438.640       88139.130 
#> Total # zero repeaters             77              135             65        
#> Percentage # zero repeaters        0.308           0.540           0.260     
#> Mean Interpurchase time            7.361           5.756           9.462     
#> (SD)                               6.791           6.394           12.266    
#> 

# add contextual factors
data("apparelStaticCov")
clv.data.apparel.cov <-
 SetStaticCovariates(clv.data.apparel,
                     data.cov.life = apparelStaticCov,
                     names.cov.life = "Gender",
                     data.cov.trans = apparelStaticCov,
                     names.cov.trans = "Gender")

# additional info about the covariates
summary(clv.data.apparel.cov)

#> CLV Transaction Data with Static Covariates 
#>                                 
#> Time unit         Weeks         
#> Estimation length 40.0000 Weeks 
#> Holdout length    39.71429 Weeks
#> 
#> Transaction Data Summary 
#>                                    Estimation      Holdout         Total     
#> Number of customers                -               -               250       
#> First Transaction in period        2005-01-03      2005-10-11      2005-01-03
#> Last Transaction in period         2005-10-10      2006-07-16      2006-07-16
#> Total # Transactions               1311            946             2257      
#> Mean # Transactions per cust       5.244           8.226           9.028     
#> (SD)                               6.082           8.934           12.603    
#> Mean Spending per Transaction      39.436          38.519          39.051    
#> (SD)                               42.649          59.723          50.503    
#> Total Spending                     51700.490       36438.640       88139.130 
#> Total # zero repeaters             77              135             65        
#> Percentage # zero repeaters        0.308           0.540           0.260     
#> Mean Interpurchase time            7.361           5.756           9.462     
#> (SD)                               6.791           6.394           12.266    
#> 
#> Covariates                           
#> Trans. Covariates    Gender
#>        # covs        1     
#> Life.  Covariates    Gender
#>        # covs        1     
#> 

# }